% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{+}
\alias{+}
\alias{+,ngchmVersion2,ngchmAxis-method}
\title{Add an ngchmAxis to an NG-CHM}
\usage{
\S4method{+}{ngchmVersion2,ngchmAxis}(e1, e2)
}
\arguments{
\item{e1}{An NG-CHM.}

\item{e2}{An NG-CHM Axis.}
}
\description{
Returns the NG-CHM (e1) augmented with the specified NG-CHM axis (e2).
}
