% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaidy.R
\name{$.shaidyRepo}
\alias{$.shaidyRepo}
\title{Provide a simpler method for accessing repo methods}
\usage{
\method{$}{shaidyRepo}(repo, method)
}
\arguments{
\item{repo}{The repository to obtain the method for}

\item{method}{The name of the method to obtain}
}
\value{
A function that calls the method with the repository as its first parameter
}
\description{
Provide a simpler method for accessing repo methods
}
