% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{castAsInteger}
\alias{castAsInteger}
\title{Helper function to cast variables as integers.}
\usage{
castAsInteger(variableToCast)
}
\arguments{
\item{variableToCast}{Variable to cast as integer}
}
\value{
integer value of variableToCast
}
\description{
If variable value is far from integer, print error message and stop.
}
