% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{castListAsInteger}
\alias{castListAsInteger}
\title{Helper function to cast list as integer}
\usage{
castListAsInteger(listToCast)
}
\arguments{
\item{listToCast}{List to cast as integer}
}
\value{
list with members cast to integers
}
\description{
If variable value is far from integer, print error message and stop.
}
