% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\docType{methods}
\name{chmAdd}
\alias{chmAdd}
\alias{chmAdd,ngchm-method}
\title{Add a list of objects to a NGCHM.}
\usage{
chmAdd(chm, ...)

\S4method{chmAdd}{ngchm}(chm, ...)
}
\arguments{
\item{chm}{The chm to add the object(s) to.}

\item{...}{Zero or more objects to add to the NGCHM.}
}
\value{
The extended chm.
}
\description{
Each additional parameter is added to the NGCHM according to its type.  Objects that
require additional information (such as an axis) cannot be added using this function.
Objects that can be added are layers (including numeric matrices), datasets, and colormaps.
}
\seealso{
"chmAddAxisType"

"chmAddColormap"

"chmAddDataset"

"chmAddLayer"

"chmAddMetaData"
}

