% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\docType{methods}
\name{chmAddClassBar}
\alias{chmAddClassBar}
\alias{chmAddClassBar,ngchm,character,ngchmBar-method}
\title{Add a classification bar to a NGCHM.}
\usage{
chmAddClassBar(chm, where, bar)

\S4method{chmAddClassBar}{ngchm,character,ngchmBar}(chm, where, bar)
}
\arguments{
\item{chm}{The chm to add the classification bar to.}

\item{where}{The chm axis(axes) to add the classification bar to. Must be one of "row", "column", or "both".}

\item{bar}{The classification bar to add to the chm.}
}
\value{
The extended chm.
}
\description{
This function is deprecated and will be removed in a future version.  Please use
chmAddCovariateBar instead.
Add a classification bar to a Next Generation Clustered Heat Map (NGCHM) and
return the extended CHM.
}
\seealso{
chmNewCovariateBar

chmAddCovariateBar

chmNewClassBar

ngchmBar-class
}

