% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\docType{methods}
\name{chmAddCovariateBar}
\alias{chmAddCovariateBar}
\alias{chmAddCovariateBar,ngchm,character,list-method}
\alias{chmAddCovariateBar,ngchm,character,ngchmBar-method}
\alias{chmAddCovariateBar,ngchm,character,ngchmCovariate-method}
\alias{chmAddCovariateBar,ngchm,character,ngchmCovariateBar-method}
\title{Add a covariate bar to a NGCHM.}
\usage{
chmAddCovariateBar(chm, where, covar, ...)

\S4method{chmAddCovariateBar}{ngchm,character,ngchmBar}(chm, where, covar)

\S4method{chmAddCovariateBar}{ngchm,character,ngchmCovariate}(chm, where, covar,
  display = "visible", thickness = as.integer(10), merge)

\S4method{chmAddCovariateBar}{ngchm,character,list}(chm, where, covar,
  display = "visible", thickness = as.integer(10), merge = NULL)
}
\arguments{
\item{chm}{The chm to add the covariate bar to.}

\item{where}{The chm axis(axes) to add the covariate bar to. Must be one of "row", "column", or "both".}

\item{covar}{The covariate or covariate bar to add to the chm.}

\item{...}{Additional parameters passed to chmNewCovariateBar if covar is a covariate.}

\item{display}{Whether the bar is displayed or not. Defaults to "visible".}

\item{thickness}{Initial thickness of bar in pixels. Defaults to 10.}

\item{merge}{Algorithm for merging multiple items into a single pixel. Default.}
}
\value{
The extended chm.
}
\description{
Add a covariate bar to a Next Generation Clustered Heat Map (NGCHM) and
return the extended CHM.  If passed a covariate, a covariate bar will be created
(using any optional parameters supplied) and added.
}
\seealso{
chmNewCovariate

chmNewCovariateBar

ngchmCovariate-class
}

