% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmAddCSS}
\alias{chmAddCSS}
\alias{chmAddCSS,ngchm,character-method}
\alias{chmAddCSS,ngchm,character,character-method}
\title{Add custom CSS to a NGCHM.}
\usage{
chmAddCSS(chm, css)

\S4method{chmAddCSS}{ngchm,character}(chm, css)
}
\arguments{
\item{chm}{The chm to add the CSS to.}

\item{css}{The css selector and style information.}
}
\value{
The extended chm.
}
\description{
Add custom Cascading Style Sheet (CSS) to a Next Generation Clustered Heat Map (NGCHM) and
return the extended CHM.
}
\seealso{
\linkS4class{ngchmCSS}
}
