% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\docType{methods}
\name{chmAddDialog}
\alias{chmAddDialog}
\alias{chmAddDialog,ngchm,ngchmDialog-method}
\title{Add an extra dialog to a NGCHM.}
\usage{
chmAddDialog(chm, dialog)

\S4method{chmAddDialog}{ngchm,ngchmDialog}(chm, dialog)
}
\arguments{
\item{chm}{The chm to add the dialog to.}

\item{dialog}{The dialog to add to the chm.}
}
\value{
The extended chm.
}
\description{
Add an extra dialog to a Next Generation Clustered Heat Map (NGCHM) and return the extended CHM.
}
\seealso{
chmNewDialog

ngchmDialog-class
}

