% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmAddMetaData}
\alias{chmAddMetaData}
\alias{chmAddMetaData,ngchm,character,character,character-method}
\title{Add meta data.}
\usage{
chmAddMetaData(chm, where, type, value)

\S4method{chmAddMetaData}{ngchm,character,character,character}(chm, where, type, value)
}
\arguments{
\item{chm}{The chm to add the meta data to.}

\item{where}{The axis of the chm to add the meta data to (row, column, or both).}

\item{type}{The type of the meta data.}

\item{value}{A character vector or NULL.
If value is a character vector, elements of the vector will be attached as meta data to to NGCHM row of the same name.}
}
\description{
Add meta data to the rows/columns of a Next Generation Clustered Heat Map.
}
