% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\docType{methods}
\name{chmAddOverview}
\alias{chmAddOverview}
\alias{chmAddOverview,ngchm,character,numeric,numeric-method}
\alias{chmAddOverview,ngchm,character,optNumeric,optNumeric-method}
\title{Generate an overview image of the NGCHM when making it.}
\usage{
chmAddOverview(chm, format, width, height)

\S4method{chmAddOverview}{ngchm,character,optNumeric,optNumeric}(chm, format,
  width, height)
}
\arguments{
\item{chm}{The chm to add the overview to.}

\item{format}{The format of the overview ('pdf', 'png', or 'svg').}

\item{width}{The width of the overview.}

\item{height}{The height of the overview.}
}
\value{
The extended chm.
}
\description{
Generate an overview image of the NGCHM when making it.  By default, the system generates no default overview
images.  If only one of width or height is specified, the other is calculated based on the aspect ratio of
the map.
}

