% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmAddPanel}
\alias{chmAddPanel}
\alias{chmAddPanel,ngchm,panel-method}
\title{Add a panel to a NGCHM.}
\usage{
chmAddPanel(chm, panel)

\S4method{chmAddPanel}{ngchm,panel}(chm, panel)
}
\arguments{
\item{chm}{An object of class "ngchm".}

\item{panel}{An object of class "panel" to be added to the ngchm object. (e.g. a detailMap or summaryMap)}
}
\value{
The extended chm.

An updated ngchm object with the new panel added.
}
\description{
This method adds a panel to an existing ngchm object.
}
\examples{
# create a detail map pane and add it to an existing ngchm object
detailMap <- detailMap(id = "pane1")
chm <- chmAddPanel(chm, detailMap)
}
