% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmAddPanels}
\alias{chmAddPanels}
\alias{chmAddPanels,ngchm,list-method}
\title{Add a list of panels to a NGCHM.}
\usage{
chmAddPanels(chm, panels)

\S4method{chmAddPanels}{ngchm,list}(chm, panels)
}
\arguments{
\item{chm}{An object of class "ngchm".}

\item{panels}{A list of panel objects to be added to the ngchm object.}
}
\value{
The extended chm.

An updated ngchm object with the new panels added.
}
\description{
This method adds multiple panels to an existing ngchm object.
}
\examples{
# create a detail map and a summary map and add them to an existing ngchm object
detailMap <- detailMap(id = "pane1")
summaryMap <- summaryMap(id = "pane2")
chm <- chmAddPanels(chm, list(detailMap, summaryMap))
}
