\name{chmAddProperty}
\alias{chmAddProperty}
\alias{chmAddProperty,ngchm,character,character-method}
\title{Add custom property to a NGCHM.}
\usage{
  chmAddProperty(chm, label, value)
}
\arguments{
  \item{chm}{The chm to add the property to.}

  \item{label}{The property label.}

  \item{value}{The property value.}
}
\value{
  The extended chm.
}
\description{
  Add custom property to a Next Generation Clustered Heat
  Map (NGCHM) and return the extended CHM.
}
\seealso{
  ngchmProperty-class
}

