\name{chmAddTag}
\alias{chmAddTag}
\alias{chmAddTag,ngchm,character,character-method}
\title{Add tags to a NGCHM.}
\usage{
  chmAddTag(chm, tag)
}
\arguments{
  \item{chm}{The chm to add the dataset to.}

  \item{tag}{The tag(s) to add to the chm.}
}
\value{
  The extended chm.
}
\description{
  Add one or more tags to a Next Generation Clustered Heat
  Map (NGCHM) and return the extended CHM.
}

