\name{chmAddTemplate}
\alias{chmAddTemplate}
\alias{chmAddTemplate,ngchm,character,character,optList-method}
\title{Add a file template to the NGCHM.}
\arguments{
  \item{chm}{The chm to add the file template to.}

  \item{source.path}{A string giving the path to the
  template.}

  \item{dest.path}{A string giving the relative path where
  to store the template in the generated CHM.}

  \item{substitutions}{A list (may be empty) and
  substitutions to make in the template.}
}
\value{
  The extended chm.
}
\description{
  Add a file template to the NGCHM.
}

