\name{chmAddToolbox}
\alias{chmAddToolbox}
\alias{chmAddToolbox,ngchm,character,character,character,character-method}
\title{Add standard toolbox to an axis}
\usage{
  chmAddToolbox(CHM, axis, axistype, datasetname, idstr)
}
\arguments{
  \item{CHM}{CHM to which the toolbox will be added}

  \item{axis}{the axis to which the toolbix will be added}

  \item{axistype}{the type of the axis}

  \item{datasetname}{the name of the dataset to use}

  \item{idstr}{string to append to toolbox menu labels
  (default '')}
}
\description{
  Add a standard toolbox to an axis
}

