\name{chmAddToolbox2}
\alias{chmAddToolbox2}
\alias{chmAddToolbox2,ngchm,character,character-method}
\title{Add standard toolbox to an element}
\usage{
  chmAddToolbox2(CHM, datasetname, idstr)
}
\arguments{
  \item{CHM}{CHM to which the toolbox will be added}

  \item{datasetname}{the name of the dataset to use}

  \item{idstr}{string to append to toolbox menu labels
  (default '')}
}
\description{
  Add a standard toolbox to an element
}

