% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmAddToolboxR2}
\alias{chmAddToolboxR2}
\alias{chmAddToolboxR2,ngchm,character,character,character-method}
\title{Add Toolbox R2 to NG-CHM}
\usage{
chmAddToolboxR2(CHM, axistype, datasetname, idstr)

\S4method{chmAddToolboxR2}{ngchm,character,character,character}(CHM, axistype, datasetname, idstr)
}
\arguments{
\item{CHM}{An object of class 'ngchm'.}

\item{axistype}{A single character string specifying the type of the axis.}

\item{datasetname}{A single character string specifying the name of the dataset.}

\item{idstr}{string to append to toolbox menu labels (default '')}
}
\value{
An updated 'ngchm' object with the new toolbox of type R2 added.
}
\description{
This function adds a toolbox of type R2 to a NG-CHM (Next-Generation Clustered Heat Map) object.
}
