% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmAddToolboxRC}
\alias{chmAddToolboxRC}
\alias{chmAddToolboxRC,ngchm,character,character,character,character-method}
\alias{chmAddToolboxRC,ngchm,character,character,character-method}
\title{Add Toolbox RC to NG-CHM}
\usage{
chmAddToolboxRC(CHM, rowtype, coltype, datasetname, idstr)

\S4method{chmAddToolboxRC}{ngchm,character,character,character,character}(CHM, rowtype, coltype, datasetname, idstr)
}
\arguments{
\item{CHM}{An object of class 'ngchm'.}

\item{rowtype}{A single character string specifying the type of the row.}

\item{coltype}{A single character string specifying the type of the column.}

\item{datasetname}{A single character string specifying the name of the dataset.}

\item{idstr}{string to append to toolbox menu labels (default '')}
}
\value{
An updated 'ngchm' object with the new toolbox of type RC added.
}
\description{
This function adds a toolbox of type RC to a NG-CHM (Next-Generation Clustered Heat Map) object.
}
