% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmAddUMAP}
\alias{chmAddUMAP}
\title{Add UMAP coordinates to an NG-CHM.}
\usage{
chmAddUMAP(hm, axis, umap, basename = "UMAP")
}
\arguments{
\item{hm}{The NGCHM to add the coordinates to.}

\item{axis}{The NGCHM axis ("row" or "column") to add the coordinates to.}

\item{umap}{TSNE coordinates (output of \code{\link[umap:umap]{umap::umap()}}) for the specified NGCHM axis.}

\item{baseName}{The prefix to use for the coordinate names.}
}
\value{
The NGCHM with added coordinates.
}
\description{
Add UMAP coordinates as hidden covariate bars to an axis of an NG-CHM.  One hidden
covariate bar is added for each UMAP coordinate.  Coordinates have names 'BASENAME.coordinate.N',
where BASENAME is specified by the parameter basename (default UMAP) and N ranges from 1 to the number of
added covariate bars.
}
\examples{
data(TCGA.GBM.EXPR, package='NGCHMDemoData');
umc <- umap::umap(t(TCGA.GBM.EXPR));
hm <- chmNew ("gbm", TCGA.GBM.EXPR);
hm <- chmAddUMAP(hm, "column", umc);

}
\seealso{
\code{\link[=chmAddPCA]{chmAddPCA()}}

\code{\link[=chmAddTSNE]{chmAddTSNE()}}

\code{\link[=chmAddUWOT]{chmAddUWOT()}}

\code{\link[=chmAddReducedDim]{chmAddReducedDim()}}
}
