% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmAxis}
\alias{chmAxis}
\title{Create a new Axis for adding to an NG-CHM.}
\usage{
chmAxis(axis, ...)
}
\arguments{
\item{axis}{The name of the axis}

\item{...}{Objects to add to the axis}
}
\value{
An object of class ngchmAxis
}
\description{
This function creates a new Axis for adding to a Next Generation Clustered Heat Map.  You can specify any
axis name here, but chmAdd only accepts row, column, and both.
}
\seealso{
\code{\link[=chmAdd]{chmAdd()}}
}
