% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmBrowse}
\alias{chmBrowse}
\title{Browse the NGCHMs on the specified server in the viewer.}
\usage{
chmBrowse(server = NULL, viewer = NULL)
}
\arguments{
\item{server}{The NG-CHM server to be browsed. If NULL, the function will use the first server in the list of available servers.}

\item{viewer}{The function to be used to open the web browser. If NULL, the function will use the 'browseURL' function.}
}
\value{
None. This function is used for its side effects of opening a web browser to view the NG-CHM server.
}
\description{
Opens the NG-CHM browser page in the viewer.
}
\seealso{
\code{\link[utils:browseURL]{utils::browseURL()}}
}
