% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmColors}
\alias{chmColors}
\title{Get the colors of an ngchmColormap, ngchmLayer, ngchmBar, or ngchmCovariate.}
\usage{
chmColors(x)
}
\arguments{
\item{x}{The object to get the colors of.}
}
\value{
A character string vector of the map colors.
}
\description{
Get the colors of an ngchmColormap, ngchmLayer, ngchmBar, or ngchmCovariate.
}
\examples{
data(TCGA.GBM.EXPR, package='NGCHMDemoData');
chmColors (chmNewDataLayer('New Layer', TCGA.GBM.EXPR[1:50,1:50))

}
\seealso{
\linkS4class{ngchm}
}
