% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{chmCreateServer}
\alias{chmCreateServer}
\title{Create an ngchmServer object for an MDS2 deployment server directory.}
\usage{
chmCreateServer(servername, cfgDir = NULL, cfgServer = NULL,
  theJarFile = NULL, serverURL = NULL)
}
\arguments{
\item{servername}{The name of the new server object.}

\item{cfgDir}{The absolute path of the MDS2 directory on the deployServer.}

\item{cfgServer}{The server on which cfgDir is located.}
}
\description{
Create an ngchmServer object called 'servername' for the MDS2 NGCHM deployment directory 'cfgDir' on
server 'deployServer'.  If deployServer is NULL (default), 'cfgDir' is assummed to be on the local machine.
}

