% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{chmCreateServerProtocol}
\alias{chmCreateServerProtocol}
\title{Create and register an NGCHM server protocol implementation.}
\usage{
chmCreateServerProtocol(protocolName, paramValidator, installMethod,
  uninstallMethod, makePrivate, makePublic)
}
\arguments{
\item{protocolName}{The name of this protocol implementation.}

\item{paramValidator}{A function(list) for validating a list of optional parameters specified for a new server.}

\item{installMethod}{A function(server,chm) for installing an NG-CHM.}

\item{uninstallMethod}{A function(server,chmname) for uninstalling an NG-CHM.}

\item{makePrivate}{A function(server,chmname) for hiding an NG-CHM.}

\item{makePublic}{A function(server,chmname) for showing an NG-CHM.}
}
\description{
This function creates and registers a protocol implementation for manipulating
an NGCHM server.
}

