% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{chmCurrentCollection}
\alias{chmCurrentCollection}
\title{Get the user's current collection}
\usage{
chmCurrentCollection()
}
\value{
the identity of the current collection
}
\description{
Get the user's current collection
}
\seealso{
\code{\link[=chmSetCollection]{chmSetCollection()}}
}
