% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{chmCurrentServer}
\alias{chmCurrentServer}
\title{Get the user's current server}
\usage{
chmCurrentServer()
}
\value{
the identity of the current server
}
\description{
Get the user's current server
}

