% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmDefaultColOrder}
\alias{chmDefaultColOrder}
\title{Return default column order of an NGCHM}
\usage{
chmDefaultColOrder(chm)
}
\arguments{
\item{chm}{An NGCHM containing at least one layer}
}
\value{
Shaid of a dendrogram suitable for use as the
chm's column order.
}
\description{
Return default column order of an NGCHM
}
