% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{chmDefaultRowOrder}
\alias{chmDefaultRowOrder}
\title{Return default row order of an NGCHM}
\usage{
chmDefaultRowOrder(chm)
}
\arguments{
\item{chm}{An NGCHM containing at least one layer}
}
\value{
Shaid of a dendrogram suitable for use as the
        chm's row order.
}
\description{
Return default row order of an NGCHM
}

