% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\docType{methods}
\name{chmDeployDir}
\alias{chmDeployDir}
\alias{chmDeployDir,ngchmServer-method}
\title{Get the deploy directory of a NGCHM server.}
\usage{
chmDeployDir(server)

\S4method{chmDeployDir}{ngchmServer}(server)
}
\arguments{
\item{server}{The server whose deploy directory is required.}
}
\value{
The path on the server to the deploy directory.
}
\description{
Return the name of the directory on a Next Generation Clustered Heat Map (NGCHM) server into
which a NGCHM must be copied for installation.
}
\seealso{
ngchmServer-class
}

