% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmExportToFile}
\alias{chmExportToFile}
\title{Export a standalone NGCHM to a file.}
\usage{
chmExportToFile(
  chm,
  filename,
  overwrite = FALSE,
  shaidyMapGen,
  shaidyMapGenJava,
  shaidyMapGenArgs
)
}
\arguments{
\item{chm}{The NGCHM to export}

\item{filename}{The file in which to save the rendered NGCHM}

\item{overwrite}{Overwrite file iff true (default false)}

\item{shaidyMapGen}{Path to shaidyMapGen jar file (default to value of environment variable SHAIDYMAPGEN)}

\item{shaidyMapGenJava}{Path to java executable with which to run shaidyMapGen (default to value of environment variable SHAIDYMAPGENJAVA or java)}

\item{shaidyMapGenArgs}{Additional arguments to pass to java when running shaidyMapGen (default to value of environment variable SHAIDYMAPGENARGS)}
}
\value{
the rendered NGCHM
}
\description{
Create a standalone viewer for the NGCHM in the specified file.
}
