% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmGetDataset}
\alias{chmGetDataset}
\alias{chmGetDataset,ngchmLayer-method}
\title{Get the dataset from an NG-CHM object}
\usage{
chmGetDataset(object)

\S4method{chmGetDataset}{ngchmLayer}(object)
}
\arguments{
\item{object}{An NG-CHM object containing an ngchmDataset}
}
\value{
The dataset associated with the specified object.
}
\description{
This function retrieves the dataset associated with a specific NG-CHM
(Next-Generation Clustered Heat Map).
}
