% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{chmGetDeployServerConfig}
\alias{chmGetDeployServerConfig}
\title{Get per-user configuration for a specific deploy Server.}
\usage{
chmGetDeployServerConfig(server)
}
\arguments{
\item{server}{The server for which the configuration is to be retrieved.
This can be either a character string representing the server name or an object
of class 'ngchmServer'.}
}
\value{
The configuration of the specified server if it exists, otherwise NULL.
}
\description{
This function retrieves the configuration of a specified NG-CHM
(Next-Generation Clustered Heat Map) deployment server.
}
\examples{
# Get the configuration of the server named 'myServer'.
config <- chmGetDeployServerConfig('myServer')
# Get the configuration of the server represented by the 'myServer' object.
config <- chmGetDeployServerConfig(myServer)
}
