% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmGetProperty}
\alias{chmGetProperty}
\alias{chmGetProperty,ngchmVersion2,character-method}
\alias{chmGetProperty,ngchmVersion2-method}
\title{Get Property from NG-CHM}
\usage{
chmGetProperty(object, label)

\S4method{chmGetProperty}{ngchmVersion2,character}(object, label)
}
\arguments{
\item{object}{An object of class 'ngchmVersion2' representing the NG-CHM from which
the property is to be retrieved.}

\item{label}{A single character string specifying the label of the property to be retrieved.}
}
\value{
The property associated with the specified label in the 'ngchmVersion2' object.
}
\description{
This function retrieves a specific property from a NG-CHM (Next-Generation Clustered Heat Map) object.
}
