% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\docType{methods}
\name{chmGetURL}
\alias{chmGetURL}
\alias{chmGetURL,character,character-method}
\alias{chmGetURL,character,ngchm-method}
\alias{chmGetURL,ngchmServer,character-method}
\alias{chmGetURL,ngchmServer,ngchm-method}
\title{Get the URL for an installed NGCHM.}
\usage{
chmGetURL(server, chm)

\S4method{chmGetURL}{ngchmServer,character}(server, chm)

\S4method{chmGetURL}{ngchmServer,ngchm}(server, chm)

\S4method{chmGetURL}{character,ngchm}(server, chm)

\S4method{chmGetURL}{character,character}(server, chm)
}
\arguments{
\item{server}{The server on which the NGCHM is installed.}

\item{chm}{The NGCHM for which the URL is required.}
}
\description{
Return the URL for accessing the specified Next Generation Clustered Heat Map (NGCHM) on the specified server.
}
\seealso{
ngchmServer-class

ngchm-class
}

