% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmLabel<-}
\alias{chmLabel<-}
\title{Set the label/name of an NG-CHM object}
\usage{
chmLabel(x) <- value
}
\arguments{
\item{x}{The NG-CHM object on which to set the label/name.}

\item{value}{The new name (a single character string).}
}
\value{
The modified NG-CHM object.
}
\description{
Set the label/name of an NG-CHM object
}
\examples{
hm <- chmNew("Old name")
chmLabel(hm) <- "A new name"

}
\seealso{
\link{chmLabel}
}
