% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmListFunctions}
\alias{chmListFunctions}
\title{List the predefined Javascript functions available for use in NGCHM menus.}
\usage{
chmListFunctions(re = ".*")
}
\arguments{
\item{re}{The regular expression to match. This should be a single character string.
Default is ".*", which matches all functions.}
}
\value{
None. This function is used for its side effects of printing the names and descriptions of the matching functions.
}
\description{
This function lists the predefined Javascript functions available for use in NGCHM menus.
}
\examples{
chmListFunctions() # List all functions.
chmListFunctions('^chm') # List all functions whose names start with 'chm'.
}
\seealso{
\code{\link[=chmAddMenuItem]{chmAddMenuItem()}}

\code{\link[=chmGetFunction]{chmGetFunction()}}

\code{\link[=chmRegisterFunction]{chmRegisterFunction()}}

\code{\link[=grep]{grep()}}
}
