% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmListServers}
\alias{chmListServers}
\title{Return the names of the NGCHM servers defined to date in this session.}
\usage{
chmListServers()
}
\description{
Return the names of the NGCHM servers defined to date in this session.
}
