% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmLoadCHM}
\alias{chmLoadCHM}
\alias{chmLoadCHM,ngchmServer,character-method}
\alias{chmLoadCHM,character,character-method}
\alias{chmLoadCHM,character,missing-method}
\title{Load CHM from NG-CHM server}
\usage{
chmLoadCHM(serverOrURL, name)

\S4method{chmLoadCHM}{ngchmServer,character}(serverOrURL, name)

\S4method{chmLoadCHM}{character,character}(serverOrURL, name)

\S4method{chmLoadCHM}{character,missing}(serverOrURL, name)
}
\arguments{
\item{serverOrURL}{An object of class 'ngchmServer' representing the server from which the NG-CHM
is to be loaded.}

\item{name}{A single character string specifying the name of the NG-CHM to be loaded.}
}
\value{
An object of class 'ngchm' representing the loaded NG-CHM.
}
\description{
Load an R CHM object from an NG-CHM server.  The CHM concerned must have been built
using this library, version 0.9.4 or later.
}
