% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chmFromJSON.R
\name{chmLoadShaidyCHM}
\alias{chmLoadShaidyCHM}
\title{Load an NG-CHM from an NG-CHM server.}
\usage{
chmLoadShaidyCHM(mapid, debug = FALSE)
}
\arguments{
\item{mapid}{An NG-CHM ShaidyID that identifies the NG-CHM to download.}

\item{debug}{If TRUE, return a list containing additional information .}
}
\value{
An object of class ngchm.
}
\description{
Load an NG-CHM from an NG-CHM server.
}
\seealso{
\code{\link[=chmInstall]{chmInstall()}}

\code{\link[=ngchmPushSourceServer]{ngchmPushSourceServer()}}
}
