% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmMake}
\alias{chmMake}
\alias{chmMake,ngchm-method}
\title{Compile a NGCHM.}
\usage{
chmMake(chm, ...)

\S4method{chmMake}{ngchm}(chm, ...)
}
\arguments{
\item{chm}{An object of class "ngchm".}

\item{...}{Additional arguments.}
}
\value{
The chm

A modified ngchm object with updated row and column orders.
}
\description{
Deprecated.  Users should no longer call this method directly.

This method prepares an ngchm object for further processing by ensuring that the row and
column orders are in the correct format.
It computes row and column orders if required, and converts different types of row and column orders to SHAIDs.
}
\details{
Compiles the specified Next Generation Clustered Heat Map (NGCHM) in preparation for installation.
}
\seealso{
\linkS4class{ngchmServer}

\linkS4class{ngchm}

\code{\link[=chmNew]{chmNew()}}

\code{\link[=chmInstall]{chmInstall()}}

\code{\link[=ngchmMakeFormat.original]{ngchmMakeFormat.original()}}
}
