% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmMakePrivate}
\alias{chmMakePrivate}
\alias{chmMakePrivate,ngchmServer,character-method}
\alias{chmMakePrivate,ngchmServer,ngchm-method}
\alias{chmMakePrivate,character,ngchm-method}
\alias{chmMakePrivate,character,character-method}
\title{Make an installed NGCHM private.}
\usage{
chmMakePrivate(server, chm)

\S4method{chmMakePrivate}{ngchmServer,character}(server, chm)

\S4method{chmMakePrivate}{ngchmServer,ngchm}(server, chm)

\S4method{chmMakePrivate}{character,ngchm}(server, chm)

\S4method{chmMakePrivate}{character,character}(server, chm)
}
\arguments{
\item{server}{The server on which the NGCHM to make private is installed.}

\item{chm}{The NGCHM to make private.}
}
\description{
Make private the specified Next Generation Clustered Heat Map (NGCHM) on the specified server.
}
\seealso{
\linkS4class{ngchmServer}

\linkS4class{ngchm}

\code{\link[=chmInstall]{chmInstall()}}

\code{\link[=chmUninstall]{chmUninstall()}}

\code{\link[=chmMakePublic]{chmMakePublic()}}
}
