% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmMakePublic}
\alias{chmMakePublic}
\alias{chmMakePublic,ngchmServer,character-method}
\alias{chmMakePublic,ngchmServer,ngchm-method}
\alias{chmMakePublic,character,ngchm-method}
\alias{chmMakePublic,character,character-method}
\title{Make NG-CHM Public on Server}
\usage{
chmMakePublic(server, chm)

\S4method{chmMakePublic}{ngchmServer,character}(server, chm)

\S4method{chmMakePublic}{ngchmServer,ngchm}(server, chm)

\S4method{chmMakePublic}{character,ngchm}(server, chm)

\S4method{chmMakePublic}{character,character}(server, chm)
}
\arguments{
\item{server}{An object of class 'ngchmServer' representing the server where the NG-CHM is hosted.}

\item{chm}{A single character string specifying the name of the NG-CHM to be made public.}
}
\value{
No return value. The function is called for its side effect of making the specified NG-CHM public on the specified server.
}
\description{
This function makes a specific NG-CHM (Next-Generation Clustered Heat Map) public on a specified server.
}
\seealso{
\linkS4class{ngchmServer}

\linkS4class{ngchm}

\code{\link[=chmInstall]{chmInstall()}}

\code{\link[=chmUninstall]{chmUninstall()}}

\code{\link[=chmMakePrivate]{chmMakePrivate()}}
}
