\name{chmName}
\alias{chmName}
\alias{chmName,ngchm-method}
\title{Get the name of a NGCHM.}
\usage{
  chmName(chm)
}
\arguments{
  \item{chm}{The CHM for which the name is required.}
}
\value{
  A string.
}
\description{
  This function returns the name of a Next Generation
  Clustered Heat Map (NGCHM) object.
}
\seealso{
  ngchm-class
}

