% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmNewDialog}
\alias{chmNewDialog}
\title{Create a new Dialog for a NGCHM.}
\usage{
chmNewDialog(id, title, fn)
}
\arguments{
\item{id}{The html id for the dialog.}

\item{title}{The dialog title / menu entry name.}

\item{fn}{The javascript function for customizing the dialog's contents.}
}
\value{
An object of class ngchmDialog
}
\description{
This function creates a new Dialog suitable for adding to a Next Generation Clustered Heat Map.
}
\seealso{
chmAdd

chmAddDialog
}

