% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmOriginalColOrder}
\alias{chmOriginalColOrder}
\title{Return original column order of an NGCHM}
\usage{
chmOriginalColOrder(chm)
}
\arguments{
\item{chm}{An NGCHM containing at least one layer}
}
\value{
Shaid of a label order suitable for use as the
chm's column order.
}
\description{
Return original column order of an NGCHM
}
