% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{chmOriginalRowOrder}
\alias{chmOriginalRowOrder}
\title{Return original row order of an NGCHM}
\usage{
chmOriginalRowOrder(chm)
}
\arguments{
\item{chm}{An NGCHM containing at least one layer}
}
\value{
Shaid of a label order suitable for use as the
        chm's row order.
}
\description{
Return original row order of an NGCHM
}

