% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{chmRandomColOrder}
\alias{chmRandomColOrder}
\title{Return random column order of an NGCHM}
\usage{
chmRandomColOrder(chm)
}
\arguments{
\item{chm}{An NGCHM containing at least one layer}
}
\value{
Shaid of a label order suitable for use as the
        chm's column order.
}
\description{
Return random column order of an NGCHM
}

