% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmRandomRowOrder}
\alias{chmRandomRowOrder}
\title{Return random row order of an NGCHM}
\usage{
chmRandomRowOrder(chm)
}
\arguments{
\item{chm}{An NGCHM containing at least one layer}
}
\value{
Shaid of a label order suitable for use as the
chm's row order.
}
\description{
Return random row order of an NGCHM
}
