% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmRegisterAxisFunction}
\alias{chmRegisterAxisFunction}
\title{Register a predefined Javascript function for use in NGCHM Axis menus.}
\usage{
chmRegisterAxisFunction(type, label, fn)
}
\arguments{
\item{type}{The axis type required by this function.}

\item{label}{The name of the axis menu entry to be used for this function.}

\item{fn}{The Javascript function to register.}
}
\description{
This function registers a Javascript function that will be automatically
added to the appropriate axis menu(s) when building a Next Generation
Clustered Heat Map for axes that match the function's axis type.
This function is intended for use by NGCHM system developers.
}
\seealso{
\code{\link[=chmAddAxisType]{chmAddAxisType()}}

\code{\link[=chmRegisterMatrixFunction]{chmRegisterMatrixFunction()}}

\code{\link[=chmRegisterTypeMapper]{chmRegisterTypeMapper()}}

\code{\link[=chmNewFunction]{chmNewFunction()}}
}
