% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmRegisterGetMetadataFunction}
\alias{chmRegisterGetMetadataFunction}
\title{Define and register a Javascript function for obtaining a specific metadata value.}
\usage{
chmRegisterGetMetadataFunction(functionName, metadataColumnName)
}
\arguments{
\item{functionName}{A single character string specifying the name of the
function to be registered.}

\item{metadataColumnName}{A single character string specifying the name
of the metadata column to be retrieved by the function.}
}
\value{
The registered function.
}
\description{
This function defines and registers a Javascript function for obtaining a specific
metadata value and returning it as a javascript list.  The function is suitable for
use as an axis type accessor function.
}
\seealso{
\code{\link[=chmAddAxisType]{chmAddAxisType()}}

\code{\link[=chmGetFunction]{chmGetFunction()}}

\code{\link[=chmListFunctions]{chmListFunctions()}}
}
