% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{chmRegisterServer}
\alias{chmRegisterServer}
\title{Register an ngchmServer.}
\usage{
chmRegisterServer(uuid, server)
}
\arguments{
\item{uuid}{A string that identifies the server namespace.}

\item{server}{The ngchmServer to register.}
}
\description{
This function registers an ngchmServer that can be used when
making and installing a Next Generation Clustered Heat Map.
}
\seealso{
chmMake

chmInstall

chmUninstall

chmUnregisterServer

ngchmServer-class
}

