% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmRegisterToolboxFunction}
\alias{chmRegisterToolboxFunction}
\title{Register a Javascript function for use in the NGCHM toolbox.}
\usage{
chmRegisterToolboxFunction(tbtype, menulabel, jsfn)
}
\arguments{
\item{menulabel}{A single character string specifying the menu label of the
toolbox function.}

\item{jsfn}{The function to be registered. This should be an object of class 'ngchmJS'.}
}
\value{
None. This function is used for its side effects of registering a new
toolbox function.
}
\description{
This function registers a Javascript function that can included in
the toolbox of an NGCHM.  This function is
intended for use by NGCHM system developers.
}
\seealso{
\code{\link[=chmNewFunction]{chmNewFunction()}}

\linkS4class{ngchmAxisFunction}

\linkS4class{ngchmMatrixFunction}
}
