% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{chmRegisterType}
\alias{chmRegisterType}
\title{Register a type name.}
\usage{
chmRegisterType(typename, description)
}
\arguments{
\item{typename}{The name of the type.}

\item{description}{A description of the type.}
}
\description{
This function registers a type name used for determining row and column
linkouts.
This function is intended to be used by NGCHM system developers to record
basic information about the semantic interpretation of a type name.  Registration
of a typename is (currently) not required in order to use it.
}
\seealso{
chmListTypes

chmGetTypeInfo

chmRegisterTypeMapper
}

