% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{chmRegisterTypeMapper}
\alias{chmRegisterTypeMapper}
\title{Register a predefined Javascript function for converting values from
one type to another.}
\usage{
chmRegisterTypeMapper(fromtype, totype, fn)
}
\arguments{
\item{fromtype}{The type of values the function expects as input.}

\item{totype}{The type of values the function will produce.  The length of
totype must be shorter than fromtype.}

\item{fn}{The Javascript function to register.}
}
\description{
This function registers a Javascript function that will be automatically
added to a Next Generation Clustered Heat Map as required for converting
values from one type into another more basic type.
This function is intended for use by NGCHM system developers.
}
\seealso{
chmAddAxisType

chmRegisterAxisFunction

chmRegisterMatrixFunction

chmNewFunction
}

