% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmRegisterTypeSplitter}
\alias{chmRegisterTypeSplitter}
\title{Define and register a Javascript function for converting a lists of type values into single values.}
\usage{
chmRegisterTypeSplitter(functionName, listtype, itemtype, separator)
}
\arguments{
\item{functionName}{A single character string specifying the name of the function
to be registered.}

\item{listtype}{A single character string specifying the type of the list to be split.}

\item{itemtype}{A single character string specifying the type of the items in the
list after splitting.}

\item{separator}{A single character string specifying the separator to be used for
splitting.}
}
\value{
None. This function is used for its side effects of registering a new type
splitter.
}
\description{
This function defines and registers a Javascript function for converting a list of type values
separated by the specified separator into the single values, and registers it as a type mapper.
}
\seealso{
\code{\link[=chmGetFunction]{chmGetFunction()}}

\code{\link[=chmListFunctions]{chmListFunctions()}}

\code{\link[=chmRegisterTypeMapper]{chmRegisterTypeMapper()}}
}
