% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmServer}
\alias{chmServer}
\title{Get a registered ngchmServer object for use in making and installing NGCHMs}
\usage{
chmServer(name)
}
\arguments{
\item{name}{The name of the ngchmServer desired.}
}
\value{
An object of class ngchmServer if found, NULL otherwise.  If multiple servers of the same
name have been defined (in different namespaces), the most recently defined is returned.
}
\description{
This function returns a ngchmServer object that can be used when
making and installing a Next Generation Clustered Heat Map.
}
\seealso{
chmInstall

chmUninstall

ngchmServer-class
}

