% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{chmSetDeployServerConfig}
\alias{chmSetDeployServerConfig}
\title{Specify per-user configuration for a specific deploy Server.}
\usage{
chmSetDeployServerConfig(server, config)
}
\arguments{
\item{server}{An object of class 'chmServer' or a character string specifying the
name of the server.}

\item{config}{A list specifying the configuration to be set for the server.}
}
\value{
None. This function is used for its side effects of setting the deployment server configuration.
}
\description{
Specify per-user configuration for a specific deploy Server.
}
