% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmSetDisplayLength}
\alias{chmSetDisplayLength}
\alias{chmSetDisplayLength,ngchmVersion2-method}
\title{Set number of characters to display for row or column labels}
\usage{
chmSetDisplayLength(chm, displayLength, rowOrCol)

\S4method{chmSetDisplayLength}{ngchmVersion2}(chm, displayLength, rowOrCol)
}
\arguments{
\item{chm}{An ngchmVersion2 object to modify}

\item{displayLength}{Numeric value to set as display length}

\item{rowOrCol}{Character string indicating which labels to modify: "row" or "col" ("column" also accepted)}
}
\value{
Modified ngchmVersion2 object with updated display length
}
\description{
Sets the maximum number of characters to display for row or column labels in the NG-CHM viewer.
}
\examples{
# Create a new NG-CHM object
chm <- chmNew("New Heat Map")
# Set row labels to display up to 20 characters
chm <- chmSetDisplayLength(chm, 20, "row")
# Set column labels to display up to 15 characters
chm <- chmSetDisplayLength(chm, 15, "col")

}
