% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{chmTreeCuts}
\alias{chmTreeCuts}
\title{Constructor function for treeCuts class}
\usage{
chmTreeCuts(numberOfCuts)
}
\description{
This allows users to use chmNew like the following:
chmNew('ngchm', rowGapLocations=chmTreeCuts(5))
}
