% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{chmTreeGaps}
\alias{chmTreeGaps}
\title{Constructor function for treeCuts class}
\usage{
chmTreeGaps(numberOfCuts)
}
\description{
This allows users to use chmNew like the following:
chmNew('ngchm', rowGapLocations=chmTreeGaps(5))
}
