\name{chmUninstall}
\alias{chmUninstall}
\alias{chmUninstall,ngchmServer,character-method}
\alias{chmUninstall,ngchmServer,ngchm-method}
\title{Remove an installed NGCHM.}
\usage{
  chmUninstall(server, chm)
}
\arguments{
  \item{server}{The server on which the NGCHM to remove is
  installed.}

  \item{chm}{The NGCHM to remove.}
}
\description{
  Remove the specified Next Generation Clustered Heat Map
  (NGCHM) from the specified server.
}
\seealso{
  ngchmServer-class

  ngchm-class

  chmInstall
}

