% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{chmUserLabelsToShaid}
\alias{chmUserLabelsToShaid}
\title{Convert user specified labels to a shaid}
\usage{
chmUserLabelsToShaid(labels)
}
\description{
Convert user specified labels to a shaid
}

