% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmWriteCustomJS}
\alias{chmWriteCustomJS}
\title{Output Javascript code required to customize an NGCHM.}
\usage{
chmWriteCustomJS(chm, filename)
}
\arguments{
\item{chm}{An object of class 'chm' representing the heat map.}

\item{filename}{A single character string specifying the name of the file
where the JavaScript will be written.}
}
\value{
None. This function is used for its side effects of writing the JavaScript to a file.
}
\description{
This function outputs the Javascript required to customize an NGCHM.
}
