% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaidy.R
\name{gitHashObject}
\alias{gitHashObject}
\title{Obtain the git hash of an existing file.}
\usage{
gitHashObject(path)
}
\arguments{
\item{path}{filename of file to hash}
}
\value{
a string containing the file hash
}
\description{
Obtain the git hash of an existing file.
}

