% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{initLogging}
\alias{initLogging}
\title{Initialize Logging}
\usage{
initLogging(log_level, log_file = NULL)
}
\arguments{
\item{log_level}{A single character string specifying the log level. This should
be one of 'TRACE', 'DEBUG', 'INFO', 'WARN', 'ERROR', or 'FATAL'.}

\item{log_file}{An optional character string specifying the name of the file where
the log will be written. If this is NULL, the log will be written to the console.}
}
\value{
None. This function is used for its side effects of initializing the logging.
}
\description{
This function initializes logging for the application.
}
