% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ngchm-class}
\alias{ngchm-class}
\title{Class representing a Next Generation Clustered Heat Map (NGCHM) under construction.}
\description{
An NG-CHM is produced by creating a heat map object with \code{\link[=chmNew]{chmNew()}}, possibly modifying or augmenting it
using additional functions, such as \code{\link[=chmAddLayer]{chmAddLayer()}}, \code{\link[=chmAddCovariateBar]{chmAddCovariateBar()}}, etc., and then
either saving it to a server using \code{\link[=chmInstall]{chmInstall()}} or saving it to a standalone file using \code{\link[=chmExportToFile]{chmExportToFile()}}.
}
\seealso{
\code{\link[=chmNew]{chmNew()}}

\code{\link[=chmRowOrder<-]{chmRowOrder<-()}}

\code{\link[=chmColOrder<-]{chmColOrder<-()}}

\code{\link[=chmAdd]{chmAdd()}}

\code{\link[=chmAddLayer]{chmAddLayer()}}

\code{\link[=chmAddCovariateBar]{chmAddCovariateBar()}}

\code{\link[=chmAddDataset]{chmAddDataset()}}

\code{\link[=chmAddAxisType]{chmAddAxisType()}}

\code{\link[=chmInstall]{chmInstall()}}

\code{\link[=chmExportToFile]{chmExportToFile()}}
}
\keyword{classes}
