% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{NGCHM-functions}
\alias{NGCHM-functions}
\title{Javascript extensions for the Next Generation Clustered Heat Map (NGCHM) Construction Library}
\description{
Currently:
\itemize{
\item Axis function View Ideogram is added for the appropriate axis types.
}
}
\seealso{
\code{\link[=chmGetFunction]{chmGetFunction()}}

\code{\link[=chmListFunctions]{chmListFunctions()}}
}
