% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmAddChmToCollection}
\alias{ngchmAddChmToCollection}
\title{Add a CHM reference to a collection}
\usage{
ngchmAddChmToCollection(collection, shaid)
}
\arguments{
\item{collection}{A list containing details of a collection}

\item{shaid}{The shaid of the CHM to add to the collection}
}
\value{
An updated list containing details of the collection
}
\description{
Add a CHM reference to a collection
}

