% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmAddCollectionToCollection}
\alias{ngchmAddCollectionToCollection}
\title{Add a collection reference to a collection}
\usage{
ngchmAddCollectionToCollection(collection, uuid)
}
\arguments{
\item{collection}{A list containing details of a collection}

\item{uuid}{The uuid of the collection to add}
}
\value{
An updated list containing details of the collection
}
\description{
The collection graph must be acyclic.
}

