% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmAddMatrixToCollection}
\alias{ngchmAddMatrixToCollection}
\title{Add a matrix reference to a collection}
\usage{
ngchmAddMatrixToCollection(collection, name, shaid)
}
\arguments{
\item{collection}{A list containing details of a collection}

\item{name}{The name to associate with the matrix reference}

\item{shaid}{The shaid of the matrix to add to the collection}
}
\value{
An updated list contaiing details of the collection
}
\description{
Add a matrix reference to a collection
}

