% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ngchmAxisType-class}
\alias{ngchmAxisType-class}
\title{Class representing a type attached to an axis in a Next Generation Clustered Heat Map (NGCHM).}
\description{
Class representing a type attached to an axis in a Next Generation Clustered Heat Map (NGCHM).
}
\keyword{classes}

