% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ngchmBar-class}
\alias{ngchmBar-class}
\title{Class representing a Covariate Bar on a Next Generation Clustered Heat Map (NGCHM).}
\description{
Class representing a Covariate Bar on a Next Generation Clustered Heat Map (NGCHM).
}
\keyword{classes}
