% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmCollectionInCollection}
\alias{ngchmCollectionInCollection}
\title{Recursively determine if collection uuid is contained in collection
A collecton always contains itself.}
\usage{
ngchmCollectionInCollection(collection, uuid)
}
\arguments{
\item{collection}{A list containing details of a collection}

\item{uuid}{A string containing the UUID to check}
}
\value{
TRUE iff collection contains uuid, otherwise FALSE
}
\description{
Recursively determine if collection uuid is contained in collection
A collecton always contains itself.
}
