% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmCollectionTree}
\alias{ngchmCollectionTree}
\title{Create a recursive description of a collection}
\usage{
ngchmCollectionTree(collection, depth = 0)
}
\arguments{
\item{collection}{A list containing details of a collection}

\item{depth}{The indentation depth to use}
}
\value{
a string vector describing the contents of the collection
}
\description{
Create a recursive description of a collection
}

