% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ngchmColormap-class}
\alias{ngchmColormap-class}
\title{Class representing a Color Map on a Next Generation Clustered Heat Map (NGCHM).}
\description{
Class representing a Color Map on a Next Generation Clustered Heat Map (NGCHM).
}
\keyword{classes}

