% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ngchmCovariate-class}
\alias{ngchmCovariate-class}
\title{Class representing a Covariate attached to a Dataset}
\description{
Class representing a Covariate attached to a Dataset
}
\keyword{classes}

