% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ngchmDataset-class}
\alias{ngchmDataset-class}
\title{Class representing a Dataset attached to a NGCHM}
\description{
Class representing a Dataset attached to a NGCHM
}
\keyword{classes}
