% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmFindRepo}
\alias{ngchmFindRepo}
\title{Find a repository, if any, that contains the requested shaid}
\usage{
ngchmFindRepo(shaid, required = TRUE)
}
\arguments{
\item{shaid}{The shaid to search for}

\item{required}{Abort if requireed and shaid not found in a known repo}
}
\value{
The first repository containing the shaid, otherwise NULL.  The
        temporary repositories are searched before source repositories.
}
\description{
Find a repository, if any, that contains the requested shaid
}

