% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmGetDataFileShaid}
\alias{ngchmGetDataFileShaid}
\title{Compute shaid for a data file}
\usage{
ngchmGetDataFileShaid(format, filename)
}
\arguments{
\item{format}{The format of the data file}

\item{filename}{The filesystem path to the data file}
}
\value{
The shaid of the data file
}
\description{
Compute shaid for a data file
}

