% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ngchmGetEnv}
\alias{ngchmGetEnv}
\title{Get the ngchm environment (for debugging only).}
\usage{
ngchmGetEnv()
}
\value{
A list representing the current environment for NG-CHM.
}
\description{
Get the library's internal ngchm environment to help debugging.
}
\examples{
ngchmGetEnv() # Get the current environment for NG-CHM.
}
