% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{ngchmGetHandleHTTR}
\alias{ngchmGetHandleHTTR}
\title{Get a HTTR handle for the server's view/WS URL}
\usage{
ngchmGetHandleHTTR(server)
}
\arguments{
\item{server}{An object of class ngchmServer}
}
\value{
An HTTR handle
}
\description{
This function returns a 'handle' suitable for use with the server's view/WS URL
}

