% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmGetLabels}
\alias{ngchmGetLabels}
\title{Get the axis labels of a shaidy dataset}
\usage{
ngchmGetLabels(shaid, axis)
}
\arguments{
\item{shaid}{The shaid of the dataset to get the labels of}

\item{axis}{The axis of the labels to get}
}
\value{
A string vector containing the axis labels of the dataset
}
\description{
Get the axis labels of a shaidy dataset
}

