% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmGetLabelsStr}
\alias{ngchmGetLabelsStr}
\title{Get the axis labels of a shaidy dataset or dendrogram}
\usage{
ngchmGetLabelsStr(shaid, axis = NULL)
}
\arguments{
\item{shaid}{The shaid of the dataset or dendrogram to get the labels of}

\item{axis}{For datasets, the axis of the labels to get}
}
\value{
A string vector containing the axis labels of the dataset or dendrogram
}
\description{
Get the axis labels of a shaidy dataset or dendrogram
}

