% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ngchmGetServerProtocol}
\alias{ngchmGetServerProtocol}
\title{Lookup a Server Protocol}
\usage{
ngchmGetServerProtocol(protocolName)
}
\arguments{
\item{protocolName}{The name of the server protocol to lookup}
}
\description{
Lookup a Server Protocol
}

