% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\docType{methods}
\name{ngchmGetShaid}
\alias{ngchmGetShaid}
\alias{ngchmGetShaid,ngchm-method}
\title{Get shaid for an object}
\usage{
ngchmGetShaid(object)

\S4method{ngchmGetShaid}{ngchm}(object)
}
\arguments{
\item{object}{The object, such as a chm, dataset, etc., for which to compute the shaid}
}
\value{
A string containing the object's shaid
}
\description{
Get shaid for an object
}

