% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmInitShaidyRepository}
\alias{ngchmInitShaidyRepository}
\title{Create a shaidy repository for NG-CHMS}
\usage{
ngchmInitShaidyRepository(shaidyDir)
}
\arguments{
\item{shaidyDir}{Basepath of local shaidy repository to create}
}
\description{
Create a shaidy repository for NG-CHMS
}

