% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ngchmJS-class}
\alias{ngchmJS-class}
\title{Class representing a custom Javascript function for a Next Generation Clustered Heat Map (NGCHM).}
\description{
Class representing a custom Javascript function for a Next Generation Clustered Heat Map (NGCHM).
}
\keyword{classes}
