% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ngchmLayer-class}
\alias{ngchmLayer-class}
\title{Class representing a Layer on a Next Generation Clustered Heat Map (NGCHM).}
\description{
Class representing a Layer on a Next Generation Clustered Heat Map (NGCHM).
}
\keyword{classes}

