% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{ngchmListServerProtocols}
\alias{ngchmListServerProtocols}
\title{List defined server protocols}
\usage{
ngchmListServerProtocols(protocolName)
}
\value{
A character vector
}
\description{
List defined server protocols
}

