% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmLoadCollection}
\alias{ngchmLoadCollection}
\title{Load a collection from a local shaidy repository}
\usage{
ngchmLoadCollection(shaidyRepo, collection.uuid = "")
}
\arguments{
\item{shaidyRepo}{The shaidy repository}

\item{collection.uuid}{A string containing the UUID of the collection to load.}
}
\value{
a list containing the collection data
}
\description{
Load a collection from a local shaidy repository
}

