% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmLoadDatasetBlob}
\alias{ngchmLoadDatasetBlob}
\title{Load a data matrix from a local shaidy repository}
\usage{
ngchmLoadDatasetBlob(shaidyRepo, shaid)
}
\arguments{
\item{shaidyRepo}{The shaidy repository}

\item{shaid}{The shaid of the dataset blob to load}
}
\value{
a list containing details of the loaded dataset
}
\description{
Load a data matrix from a local shaidy repository
}

