% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmMakeFormat.shaidy}
\alias{ngchmMakeFormat.shaidy}
\title{Make a shaidy format NGCHM.}
\usage{
ngchmMakeFormat.shaidy(chm)
}
\arguments{
\item{chm}{The shaidy format CHM to compile.}
}
\value{
The CHM
}
\description{
Make a shaidy format NGCHM.
}
