% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ngchmMatrixFunction-class}
\alias{ngchmMatrixFunction-class}
\title{Class representing a matrix function for Next Generation Clustered Heat Map (NGCHM).}
\description{
Class representing a matrix function for Next Generation Clustered Heat Map (NGCHM).
}
\keyword{classes}

