% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ngchmMetaData-class}
\alias{ngchmMetaData-class}
\title{Class representing meta data attached to an NG-CHM}
\description{
Class representing meta data attached to an NG-CHM
}
\keyword{classes}

