% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmNewCollection}
\alias{ngchmNewCollection}
\title{Create a new collection in a local shaidy repository}
\usage{
ngchmNewCollection(shaidyRepo, labels = data.frame())
}
\arguments{
\item{shaidyRepo}{The shaidy repository}

\item{labels}{Initial labels for collection (a data.frame of (Name,Value) tuples)}
}
\value{
a string containing the UUID of the newly created repository
}
\description{
Create a new collection in a local shaidy repository
}
