% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ngchmOverview-class}
\alias{ngchmOverview-class}
\title{Class representing an overview of a Next Generation Clustered Heat Map (NGCHM).}
\description{
Class representing an overview of a Next Generation Clustered Heat Map (NGCHM).
}
\keyword{classes}

