% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ngchmProperty-class}
\alias{ngchmProperty-class}
\title{Class representing a Generic Property for a Next Generation Clustered Heat Map (NGCHM).}
\description{
Class representing a Generic Property for a Next Generation Clustered Heat Map (NGCHM).
}
\keyword{classes}

