% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmPushSourceRepository}
\alias{ngchmPushSourceRepository}
\title{Push a shaidy repository onto the stack of source repositories}
\usage{
ngchmPushSourceRepository(shaidyDir, accessMethod = "file")
}
\arguments{
\item{shaidyDir}{Basepath of local shaidy repository to use as a source repository}

\item{accessMethod}{Method for accessing repository}
}
\description{
Push a shaidy repository onto the stack of source repositories
}

