% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmPushSourceServer}
\alias{ngchmPushSourceServer}
\title{Push a shaidy server onto the stack of source repositories}
\usage{
ngchmPushSourceServer(server)
}
\arguments{
\item{server}{An object of class 'ngchmServer' or a single character string
specifying the name of the server.}
}
\value{
None. This function is used for its side effects of pushing the source
server onto the Shaidy stack.
}
\description{
This function pushes a source server for NG-CHM (Next-Generation Clustered Heat Map)
onto the Shaidy stack.
}
\seealso{
\code{\link[=chmLoadShaidyCHM]{chmLoadShaidyCHM()}}

\code{\link[=chmCreateServer]{chmCreateServer()}}
}
