% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmPushTempRepository}
\alias{ngchmPushTempRepository}
\title{Push Temporary Repository for NG-CHM}
\usage{
ngchmPushTempRepository(shaidyDir)
}
\arguments{
\item{shaidyDir}{A single character string specifying the directory of the
temporary repository.}
}
\value{
None. This function is used for its side effects of pushing the temporary
repository onto the Shaidy stack.
}
\description{
This function pushes a temporary repository for NG-CHM (Next-Generation Clustered
Heat Map) onto the Shaidy stack.
}
