% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ngchmRelatedGroup-class}
\alias{ngchmRelatedGroup-class}
\title{Class representing a group of related links to a NGCHM}
\description{
Class representing a group of related links to a NGCHM
}
\keyword{classes}

