% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmRenderChm}
\alias{ngchmRenderChm}
\title{Render a shaidy NGCHM}
\usage{
ngchmRenderChm(repo, shaid)
}
\arguments{
\item{repo}{The repository containing the chm}

\item{shaid}{The shaid of the chm to render}
}
\value{
Nothing
}
\description{
Render a shaidy NGCHM
}

