% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ngchmResponseJSON}
\alias{ngchmResponseJSON}
\title{Return response content interpreted as JSON}
\usage{
ngchmResponseJSON(httrResponse)
}
\arguments{
\item{httrResponse}{The httr response object}
}
\value{
The response parsed as JSON and returned as an R object
}
\description{
Return response content interpreted as JSON
}

