% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmRowCenter}
\alias{ngchmRowCenter}
\title{Row center a shaidy dataset}
\usage{
ngchmRowCenter(shaidyRepo, shaid)
}
\arguments{
\item{shaidyRepo}{The shaidy repository}

\item{shaid}{The shaid of the dataset to row center}
}
\value{
A string containing the shaid of the row centered dataset
}
\description{
Row center a shaidy dataset
}

