% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmSaveAsDatasetBlob}
\alias{ngchmSaveAsDatasetBlob}
\title{Save a numeric matrix as a blob in a shaidy repository}
\usage{
ngchmSaveAsDatasetBlob(shaidyRepo, format, mat)
}
\arguments{
\item{shaidyRepo}{The shaidy repository}

\item{format}{The format in which to save the matrix}

\item{mat}{The data matrix}
}
\value{
The shaid of the saved blob
}
\description{
Save a numeric matrix as a blob in a shaidy repository
}
