% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmSaveAsDendrogramBlob}
\alias{ngchmSaveAsDendrogramBlob}
\title{Save a dendrogram as a blob in a shaidy repository}
\usage{
ngchmSaveAsDendrogramBlob(shaidyRepo, ddg)
}
\arguments{
\item{shaidyRepo}{The shaidy repository}

\item{ddg}{The dendrogram}
}
\value{
The shaid of the saved blob
}
\description{
Save a dendrogram as a blob in a shaidy repository
}

