% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmSaveChmAsBlob}
\alias{ngchmSaveChmAsBlob}
\title{Save an NGCHM as a shaidy blob}
\usage{
ngchmSaveChmAsBlob(shaidyRepo, chm)
}
\arguments{
\item{shaidyRepo}{The shaidy repository to write to}

\item{chm}{The NGCHM to write}
}
\value{
The shaid of the saved NGCHM
}
\description{
Save an NGCHM as a shaidy blob
}
