% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ngchmServerProtocol-class}
\alias{ngchmServerProtocol-class}
\title{Class representing a deployment method for a Next Generation Clustered Heat Map (NGCHM) server.}
\description{
Class representing a deployment method for a Next Generation Clustered Heat Map (NGCHM) server.
}
\keyword{classes}

