% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ngchmTemplate-class}
\alias{ngchmTemplate-class}
\title{Class representing a Template attached to a NGCHM}
\description{
Class representing a Template attached to a NGCHM
}
\keyword{classes}

