% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmTileDataset}
\alias{ngchmTileDataset}
\title{Get the tiles for a shaidy dataset}
\usage{
ngchmTileDataset(repo, dataset, rowOrder, colOrder)
}
\arguments{
\item{repo}{The shaidy repository in which to create the tile}

\item{dataset}{The shaid of the dataset to tile}

\item{rowOrder}{The row order of the tiles}

\item{colOrder}{The column order of the tiles}
}
\value{
a list of shaids containing the tiles
}
\description{
Get the tiles for a shaidy dataset
}

