% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ngchmTypeMapper-class}
\alias{ngchmTypeMapper-class}
\title{Class representing a type mapper function for Next Generation Clustered Heat Map (NGCHM).}
\description{
Class representing a type mapper function for Next Generation Clustered Heat Map (NGCHM).
}
\keyword{classes}

