% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ngchmUnregisterServer}
\alias{ngchmUnregisterServer}
\title{Unregister ngchmServer(s).}
\usage{
ngchmUnregisterServer(uuid, name = NULL)
}
\arguments{
\item{uuid}{A string that identifies the server namespace.}

\item{name}{The names(s) of the ngchmServer(s) to unregister.  If not specified, all ngchmServers in the
namespace are unregistered.}
}
\description{
This function unregisters one or more ngchmServer(s).
}
\seealso{
\code{\link[=ngchmRegisterServer]{ngchmRegisterServer()}}

\linkS4class{ngchmServer}
}
