% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ngchmValueProp-class}
\alias{ngchmValueProp-class}
\title{Class representing the properties of a data point in a Next Generation Clustered Heat Map (NGCHM).}
\description{
Class representing the properties of a data point in a Next Generation Clustered Heat Map (NGCHM).
}
\keyword{classes}

