% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.ngchmVersion2}
\alias{plot.ngchmVersion2}
\title{Open the NG-CHM on the specified server in the viewer.}
\usage{
\method{plot}{ngchmVersion2}(chm, server = NULL, viewer = NULL)
}
\arguments{
\item{server}{The server containing the NG-CHM.  Defaults to option "NGCHM.Server" or the first server.}

\item{viewer}{The viewer to use. Defaults to option "viewer" or browseURL.}
}
\description{
Open the NG-CHM on the specified server in the viewer.
}

