% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelClasses.R
\name{pluginPane}
\alias{pluginPane}
\title{Create a plugin pane for NG-CHM layout}
\usage{
pluginPane(id = NA, pluginName = NA)
}
\arguments{
\item{id}{Character string identifying the plugin pane. Must be specified, no default.}

\item{pluginName}{Character string specifying the plugin type (e.g., "2D ScatterPlot: UMAP (column)").
Must be specified, no default.}
}
\value{
A new \code{pluginPane} object with the specified properties
}
\description{
Creates a new plugin pane object that represents a specialized visualization
component in the NG-CHM panel layout system, such as scatter plots or bar charts.
}
\examples{
# Create a UMAP scatter plot plugin pane
plugin <- pluginPane(
  id = "pane3",
  pluginName = "2D ScatterPlot: UMAP (column)"
)

}
\seealso{
\itemize{
\item \linkS4class{pluginPane} for class details
\item \link{panel_configuration} for using plugin panes in configurations
}
}
