% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{print.ngchm.type.info}
\alias{print.ngchm.type.info}
\title{Pretty print the result returned by chmGetTypeInfo.}
\usage{
\method{print}{ngchm.type.info}(ti)
}
\description{
Pretty print the result returned by chmGetTypeInfo.
}
\seealso{
chmGetTypeInfo
}

