% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shaidy.R
\name{print.shaidyRepo}
\alias{print.shaidyRepo}
\title{Print a shaidy repository}
\usage{
\method{print}{shaidyRepo}(repo)
}
\arguments{
\item{repo}{The shaidy repository to print}
}
\value{
The shaidy repository
}
\description{
Print a shaidy repository
}

