% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{readTile}
\alias{readTile}
\title{Read a binary viewer tile as a matrix}
\usage{
readTile(filename, nrow, ncol)
}
\arguments{
\item{filename}{The name of the file to read the tile from}

\item{nrow}{Number of rows in the tile}

\item{ncol}{Number of columns in the tile}
}
\description{
Read a vector of little-endian 32-bit floats from the specified file and return
as a matrix with nrow rows and ncol columns.
}

