% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaidy.R
\name{shaidyAddFileBlob}
\alias{shaidyAddFileBlob}
\title{Add data file(s) and properties to a local shaidy repository}
\usage{
shaidyAddFileBlob(
  shaidyRepo,
  blob.type,
  blob.file,
  filename,
  properties = NULL,
  shaid = NULL
)
}
\arguments{
\item{shaidyRepo}{The shaidy repository}

\item{blob.type}{The blob.type of the data file}

\item{blob.file}{Name of the file(s) within the blob}

\item{filename}{The filesystem path(s) to the file(s) to insert}

\item{properties}{A list of additional properties to save with the file(s)}

\item{shaid}{Shaid to store the blob as.}
}
\value{
The file's shaid
}
\description{
Add data file(s) and properties to a local shaidy repository
}
