% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaidy.R
\name{shaidyBlobExists}
\alias{shaidyBlobExists}
\title{Determine if one more blobs exist in a shaidy repository}
\usage{
shaidyBlobExists(repo, shaids)
}
\arguments{
\item{repo}{The shaidy repository}

\item{shaids}{A shaid or list of shaids}
}
\value{
a boolean vector
}
\description{
Determine if one more blobs exist in a shaidy repository
}
