% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaidy.R
\name{shaidyCreateProtoBlob}
\alias{shaidyCreateProtoBlob}
\title{Create a prototype blob in a shaidy repository}
\usage{
shaidyCreateProtoBlob(shaidyRepo, blob.type)
}
\arguments{
\item{shaidyRepo}{The shaidy repository}

\item{blob.type}{The type of the prototype blob (e.g. dataset, label, chm)}
}
\value{
The file path of the prototype blob
}
\description{
Apparently a 'prototype blob' is a directory.
}
\keyword{internal}
