% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaidy.R
\name{shaidyFindRepo}
\alias{shaidyFindRepo}
\title{Find the first repository, if any, that contains the requested shaid}
\usage{
shaidyFindRepo(repos, shaid)
}
\arguments{
\item{repos}{The list of repositories to search}

\item{shaid}{The shaid to search for}
}
\value{
The first repository containing the shaid, otherwise NULL
}
\description{
Find the first repository, if any, that contains the requested shaid
}

