% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\docType{methods}
\name{shaidyGetShaid}
\alias{shaidyGetShaid}
\alias{shaidyGetShaid,ngchm-method}
\title{Get shaid for an object}
\usage{
shaidyGetShaid(object)

\S4method{shaidyGetShaid}{ngchm}(object)
}
\arguments{
\item{object}{The object, such as a chm, dataset, etc., for which to get the shaid}
}
\value{
The shaid of the object.
}
\description{
Get shaid for an object
}

