% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaidy.R
\name{shaidyHashProtoBlob}
\alias{shaidyHashProtoBlob}
\title{Compute the shaid to assign a protoblob}
\usage{
shaidyHashProtoBlob(blob.type, protoblob)
}
\arguments{
\item{blob.type}{The blob.type of the prototype blob}

\item{protoblob}{The prototype blob}
}
\value{
The shaid to assign the protoblob
}
\description{
Compute the shaid to assign a protoblob
}
