% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaidy.R
\name{shaidyInit}
\alias{shaidyInit}
\title{Initialize the shaidy subsystem}
\usage{
shaidyInit()
}
\description{
blobPath returns a function (first, ...) that accepts either a shaid or a blob type and optionally
        additional file path components and returns a filepath
}

