% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaidy.R
\name{shaidyInitRepository}
\alias{shaidyInitRepository}
\title{Create and initialize Shaidy Repository}
\usage{
shaidyInitRepository(shaidyDir, blob.types)
}
\arguments{
\item{shaidyDir}{A single character string specifying the directory where the
Shaidy repository will be initialized.}

\item{blob.types}{A character vector specifying the blob types for the Shaidy
repository.}
}
\value{
None. This function is used for its side effects of initializing the Shaidy
repository.
}
\description{
This function initializes a Shaidy repository in a specified directory with
specified blob types.
}
\examples{
shaidyInitRepository('myShaidyDir', c('blobType1', 'blobType2')) # Initialize a Shaidy repository in 'myShaidyDir' with blob types 'blobType1' and 'blobType2'.
}
