% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaidy.R
\name{shaidyLoadProvenanceDB}
\alias{shaidyLoadProvenanceDB}
\title{Load the provid -> shaid DB for a local shaidy repository}
\usage{
shaidyLoadProvenanceDB(shaidyDir)
}
\arguments{
\item{shaidyDir}{Basepath to a local shaidy repository.}
}
\value{
A shaidyProvenanceDB
}
\description{
Load the provid -> shaid DB for a local shaidy repository
}
