% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaidy.R
\name{shaidyLoadProvidDB}
\alias{shaidyLoadProvidDB}
\title{Load the provid -> labels DB for a local shaidy repository.}
\usage{
shaidyLoadProvidDB(shaidyDir)
}
\arguments{
\item{shaidyDir}{Basepath to a local shaidy repository.}
}
\value{
A shaidyProvidDB
}
\description{
Load the provid -> labels DB for a local shaidy repository.
}

