% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaidy.R
\name{shaidyLoadRepository}
\alias{shaidyLoadRepository}
\title{Load a shaidy repository}
\usage{
shaidyLoadRepository(accessMethod, shaidyDir)
}
\arguments{
\item{accessMethod}{Method for accessing repository.}

\item{shaidyDir}{Basepath to shaidy repository.}
}
\value{
A shaidyRepo
}
\description{
Load a shaidy repository
}
