% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shaidy.R
\name{shaidyNewCache}
\alias{shaidyNewCache}
\title{Create in memory shaid cache}
\usage{
shaidyNewCache(shaidyDir)
}
\arguments{
\item{shaidyDir}{Basepath to a local shaidy repository.}
}
\value{
An in memory shaid cache
}
\description{
Create in memory shaid cache
}

