% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shaidy.R
\name{shaidyProvenance}
\alias{shaidyProvenance}
\title{Create a provid from a list of label values}
\usage{
shaidyProvenance(...)
}
\arguments{
\item{...}{shaidyRepo followed by a list of name=value labels to store in the provid}
}
\value{
A string containing the provid for the list of label values.
}
\description{
Create a provid from a list of label values
}

