% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaidy.R
\name{shaidyRegisterRepoAPI}
\alias{shaidyRegisterRepoAPI}
\title{Set the methods for the repository API called api}
\usage{
shaidyRegisterRepoAPI(api, methods)
}
\arguments{
\item{api}{The name of a repository API}

\item{methods}{A list of repository methods}
}
\description{
Set the methods for the repository API called api
}

