% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaidy.R
\name{shaidyRepoAPI}
\alias{shaidyRepoAPI}
\title{Get the methods for the repository API called api}
\usage{
shaidyRepoAPI(api)
}
\arguments{
\item{api}{The name of a repository API}
}
\value{
A list of repository methods
}
\description{
Get the methods for the repository API called api
}

