% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelClasses.R
\docType{class}
\name{summaryMap-class}
\alias{summaryMap-class}
\title{summaryMap class}
\description{
This is a helper class for creating the 'panel_configuration' part of the mapConfig.json file.
This class represents a 'summaryMap' entry.
}
\section{Slots}{

\describe{
\item{\code{id}}{The ID of the summary map. This must be specified.}
}}

