% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelClasses.R
\name{summaryMap}
\alias{summaryMap}
\title{Create a summary map pane}
\usage{
summaryMap(id = NA)
}
\arguments{
\item{id}{Character string identifying the summary map pane. Must be specified, no default.}
}
\value{
A new \code{summaryMap} object with the specified ID
}
\description{
Creates a new summary map object for use in NG-CHM panel configurations.
}
\examples{
# Create a summary map pane
summary <- summaryMap(id = "pane2")

}
\seealso{
\itemize{
\item \linkS4class{summaryMap} for class details
\item \link{panel_configuration} for using summary maps in configurations
\item \link{detailMap} for detail map panes
}
}
