\docType{data}
\name{TCGA-GBM-EXPR}
\alias{TCGA-GBM-EXPR}
\title{Sample dataset containing RNA expression data from Affymetrix U133A chips
for TCGA Glioblastoma Multiforme samples.}
\description{
  Sample dataset containing RNA expression data from
  Affymetrix U133A chips for TCGA Glioblastoma Multiforme
  samples.
}
\keyword{data}
\keyword{datasets}

