% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{TCGA.GBM.EXPR}
\alias{TCGA.GBM.EXPR}
\title{Sample dataset containing RNA expression data from Affymetrix U133A chips
for TCGA Glioblastoma Multiforme samples.}
\description{
Sample dataset containing RNA expression data from Affymetrix U133A chips
for TCGA Glioblastoma Multiforme samples.
}
\examples{
data(TCGA.GBM.EXPR)
sample.type <- 'bio.tcga.barcode.sample.vial.portion.analyte.aliquot'
chm <- chmNew ('my-gbm-chm', TCGA.GBM.EXPR, rowAxisType='bio.gene.hugo',
                colAxisType=sample.type)
\dontrun{
chmMake ('my-server', chm)
chmInstall ('my-server', chm)
}
}
\keyword{data}

