% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{treeCuts-class}
\alias{treeCuts-class}
\title{Helper class for setting row/col gap locations as cuts}
\description{
Helper class for setting row/col gap locations as cuts
}
\section{Slots}{

\describe{
\item{\code{numberOfCuts}}{Integer number of cuts}
}}

