% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{validateNewLayer}
\alias{validateNewLayer}
\title{Check that the specified layer can be added to the specificed NGCHM.}
\usage{
validateNewLayer(chm, layer)
}
\arguments{
\item{chm}{The CHM to which the new layer is being added.}

\item{layer}{The new layer is being added.}
}
\description{
Stops with an error if the specified layer cannot be added to the specified NGCHM.
}

