% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{verifyNumeric}
\alias{verifyNumeric}
\title{Helper function to verify if variable is numeric.}
\usage{
verifyNumeric(variableToCheck)
}
\arguments{
\item{listToCast}{List to cast as integer}
}
\value{
list with members cast to integers
}
\description{
If not numeric, print error message and stop.
}
