% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{writeTile}
\alias{writeTile}
\title{Write a matrix as a binary viewer tile}
\usage{
writeTile(mat, filename)
}
\arguments{
\item{mat}{The matrix to output}

\item{filename}{The name of the file to write the tile to}
}
\description{
Write the matrix as a vector of 32-bit little-endian floats to the specified file.
}

