% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{+}
\alias{+}
\alias{+,ngchmVersion2,ngchmAxis-method}
\title{Add an Axis to an NG-CHM Version 2}
\usage{
\S4method{+}{ngchmVersion2,ngchmAxis}(e1, e2)
}
\arguments{
\item{e1}{An object of class 'ngchmVersion2' to which the axis is to be added.}

\item{e2}{An object of class 'ngchmAxis' representing the axis to be added.}
}
\value{
An updated 'ngchmVersion2' object with the added axis.
}
\description{
This function adds an 'ngchmAxis' to an 'ngchmVersion2' object.
}
