% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{castListAsInteger}
\alias{castListAsInteger}
\title{Helper function to cast list as integer}
\usage{
castListAsInteger(listToCast, errorMsg = NULL)
}
\arguments{
\item{listToCast}{List to cast as integer}

\item{errorMsg}{Error message if cast fails}
}
\value{
list with members cast to integers
}
\description{
If variable value is far from integer, print error message and stop.
}
