% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\docType{methods}
\name{chmAddCovariate}
\alias{chmAddCovariate}
\alias{chmAddCovariate,ngchmDataset,character,ngchmCovariate-method}
\title{Add a covariate to an auxiliary dataset.}
\usage{
chmAddCovariate(dataset, where, covariate)

\S4method{chmAddCovariate}{ngchmDataset,character,ngchmCovariate}(dataset,
  where, covariate)
}
\arguments{
\item{dataset}{The dataset to add the covariate to.}

\item{where}{The dataset axis to add the covariate to. Must be one of "row", "column", or "both".}

\item{covariate}{The covariate to add to the dataset.}
}
\value{
The extended dataset.
}
\description{
Add a covariate to an auxiliary dataset and return the extended dataset.
Do not confuse this function with the one for adding a covariate bar to an NGCHM. For
that, please refer to the function chmAddCovariateBar.
}
\seealso{
chmNewCovariate

ngchmCovariate-class
}

