\name{chmAddCSS}
\alias{chmAddCSS}
\alias{chmAddCSS,ngchm,character,character-method}
\title{Add custom CSS to a NGCHM.}
\usage{
  chmAddCSS(chm, css)
}
\arguments{
  \item{chm}{The chm to add the CSS to.}

  \item{css}{The css selector and style information.}
}
\value{
  The extended chm.
}
\description{
  Add custom Cascading Style Sheet (CSS) to a Next
  Generation Clustered Heat Map (NGCHM) and return the
  extended CHM.
}
\seealso{
  ngchmCSS-class
}

