% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmAddReducedDim}
\alias{chmAddReducedDim}
\title{Add Single Cell reduced dimension coordinates to an NG-CHM.}
\usage{
chmAddReducedDim(hm, axis, sce, dimName, maxDim, basename)
}
\arguments{
\item{hm}{The NGCHM to add the coordinates to.}

\item{axis}{The NGCHM axis ("row" or "column") to add the coordinates to.}

\item{sce}{A SingleCellExperiment or Seurat object containing the reduced dimension.}

\item{dimName}{The name of the reduced dimension to create covariate bars for.}

\item{maxDim}{The maximum number of coordinates to add (default all).}

\item{basename}{The prefix to use for the coordinate names (defaults to dimName).}
}
\value{
The NGCHM with added coordinates.
}
\description{
Add SingleCellExperiment::reducedDim or Seurat reduced dimension coordinates from a single cell object sce
as hidden covariate bars to an axis of an NG-CHM.  dimName specifies the name of the reduced
dimension of interest in sce.  One hidden covariate bar is added for each coordinate.
If specified, maxDim limits the maximum number of covariate bars added to the chm.
Coordinates have names 'BASENAME.coordinate.N', where BASENAME is specified by the parameter
basename (defaults to dimName if omitted) and N ranges from 1 to the number of added covariate bars.
}
\examples{
hm <- chmAddReducedDim(hm, "column", sce, "PCA", 3, "PC");
hm <- chmAddReducedDim(hm, "column", sce, "TSNE");

}
\seealso{
\code{\link[=chmAddPCA]{chmAddPCA()}}

\code{\link[=chmAddTSNE]{chmAddTSNE()}}

\code{\link[=chmAddUMAP]{chmAddUMAP()}}

\code{\link[=chmAddUWOT]{chmAddUWOT()}}
}
