% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmAddRelated}
\alias{chmAddRelated}
\alias{chmAddRelated,ngchm,character,character,character-method}
\title{Add a link to related information to the NGCHM.}
\usage{
chmAddRelated(chm, group, link, description)

\S4method{chmAddRelated}{ngchm,character,character,character}(chm, group, link, description)
}
\arguments{
\item{chm}{The chm to add the related link to.}

\item{group}{The name of the group this link belongs to.}

\item{link}{The link to include. Should be either an absolute URL, or a NGCHM name on the same server.}

\item{description}{A string describing the referenced link and its relationship to the current NGCHM.}
}
\value{
The extended chm.
}
\description{
Add a link to related information to the NGCHM.
}
