% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmAddTag}
\alias{chmAddTag}
\alias{chmAddTag,ngchm,character-method}
\alias{chmAddTag,ngchm,character,character-method}
\title{Add tags to a NGCHM.}
\usage{
chmAddTag(chm, tag)

\S4method{chmAddTag}{ngchm,character}(chm, tag)
}
\arguments{
\item{chm}{The chm to add the dataset to.}

\item{tag}{The tag(s) to add to the chm.}
}
\value{
The extended chm.
}
\description{
Add one or more tags to a Next Generation Clustered Heat Map (NGCHM) and return the extended CHM.
}
