% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmAddToolboxR}
\alias{chmAddToolboxR}
\alias{chmAddToolboxR,ngchm,character,character,character,character-method}
\title{Add standard toolbox to an NG-CHM axis}
\usage{
chmAddToolboxR(CHM, axis, axistype, datasetname, idstr)

\S4method{chmAddToolboxR}{ngchm,character,character,character,character}(CHM, axis, axistype, datasetname, idstr)
}
\arguments{
\item{CHM}{An object of class 'ngchm'.}

\item{axis}{A single character string specifying the axis where the toolbox will be added. Can be
"row", "column", or "both".}

\item{axistype}{A single character string specifying the type of the axis.}

\item{datasetname}{A single character string specifying the name of the dataset.}

\item{idstr}{string to append to toolbox menu labels (default '')}
}
\value{
An updated 'ngchm' object with the new toolbox added.
}
\description{
This function adds a toolbox to a NG-CHM (Next-Generation Clustered Heat Map) axis.
}
