% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\docType{methods}
\name{chmAddToolboxR2}
\alias{chmAddToolboxR2}
\alias{chmAddToolboxR2,ngchm,character,character,character-method}
\title{Add standard toolbox to an element}
\usage{
chmAddToolboxR2(CHM, axistype, datasetname, idstr)

\S4method{chmAddToolboxR2}{ngchm,character,character,character}(CHM, axistype,
  datasetname, idstr)
}
\arguments{
\item{CHM}{CHM to which the toolbox will be added}

\item{axistype}{the type of both axes}

\item{datasetname}{the name of the dataset to use}

\item{idstr}{string to append to toolbox menu labels (default '')}
}
\description{
Add a standard toolbox to an element
}

