% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\docType{methods}
\name{chmAddToolboxRC}
\alias{chmAddToolboxRC}
\alias{chmAddToolboxRC,ngchm,character,character,character,character-method}
\alias{chmAddToolboxRC,ngchm,character,character,character-method}
\title{Add standard toolbox to an element}
\usage{
chmAddToolboxRC(CHM, rowtype, coltype, datasetname, idstr)

\S4method{chmAddToolboxRC}{ngchm,character,character,character,character}(CHM,
  rowtype, coltype, datasetname, idstr)
}
\arguments{
\item{CHM}{CHM to which the toolbox will be added}

\item{rowtype}{the type of the row axis}

\item{coltype}{the type of the column axis}

\item{datasetname}{the name of the dataset to use}

\item{idstr}{string to append to toolbox menu labels (default '')}
}
\description{
Add a standard toolbox to an element
}

