% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{chmAxisType}
\alias{chmAxisType}
\title{Create a new AxisType for adding to an ngchmAxis.}
\usage{
chmAxisType(type, func)
}
\arguments{
\item{func}{A javascript function (optional) for obtaining a value of that type from the axis}

\item{axis}{The name of the axis type}
}
\value{
An object of class ngchmAxisType
}
\description{
This function creates a new AxisType for adding to an ngchmAxis.
}
\seealso{
chmAxis
}

