% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmAxisType}
\alias{chmAxisType}
\title{Create a new AxisType for adding to an ngchmAxis.}
\usage{
chmAxisType(typename, func)
}
\arguments{
\item{typename}{The name of the axis type}

\item{func}{A javascript function (optional) for obtaining a value of that type from the axis}
}
\value{
An object of class ngchmAxisType
}
\description{
This function creates a new AxisType for adding to an ngchmAxis.
}
\seealso{
chmAxis
}

