% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmBrowse}
\alias{chmBrowse}
\title{Browse the NGCHMs on the specified server in the viewer.}
\usage{
chmBrowse(server = NULL, viewer = NULL)
}
\arguments{
\item{server}{The server to browse. Defaults to option "NGCHM.Server" or the first server.}

\item{viewer}{The viewer to use. Defaults to option "viewer" or browseURL.}
}
\description{
Opens the NG-CHM browser page in the viewer.
}
\seealso{
\code{\link[utils:browseURL]{utils::browseURL()}}
}
