% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmColOrder<-}
\alias{chmColOrder<-}
\alias{chmColOrder<-,ngchm,optDendrogram-method}
\title{Set the column order of data shown in a NGCHM.}
\usage{
chmColOrder(chm) <- value

\S4method{chmColOrder}{ngchm,optDendrogram}(chm) <- value
}
\arguments{
\item{chm}{An object of class 'ngchm'.}

\item{value}{An object of class 'optDendrogram' or 'file' specifying the new column order.
If value is NULL, the labels will be displayed in the same order they are found in the first data layer.
If value is a character vector, the labels will be displayed in that order.
If value is a dendrogram, the labels displayed in the order they occur in a depth first traversal of the tree.}
}
\value{
An updated 'ngchm' object with the new column order.
}
\description{
This function sets the column order for a NG-CHM (Next-Generation Clustered Heat Map) object.
}
\seealso{
"chmRowOrder<-"
}
