% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmColorMap}
\alias{chmColorMap}
\title{Get the color map of an NG-CHM object.}
\usage{
chmColorMap(x)
}
\arguments{
\item{x}{The NG-CHM object to get the color map of.  Can be:
\itemize{
\item An object of class ngchmLayer
\item An object of class ngchmBar
\item An object of class ngchmCovariate
}}
}
\value{
An ngchmColormap
}
\description{
Get the color map of an NG-CHM object.
}
\examples{
data(TCGA.GBM.EXPR, package='NGCHMDemoData');
chmColorMap (chmNewDataLayer('New layer', TCGA.GBM.EXPR[1:3,1:3]))

}
\seealso{
\link{chmNewColorMap}
}
