% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmColors<-}
\alias{chmColors<-}
\title{Set the colors of an ngchmColormap, ngchmLayer, ngchmBar, or ngchmCovariate.}
\usage{
chmColors(x) <- value
}
\arguments{
\item{x}{The NG-CHM object on which to set the colors.}

\item{value}{A character string vector of colors.    The vector length must equal
the number of data points in the color map.}
}
\value{
The modified NG-CHM object.
}
\description{
Set the colors of an ngchmColormap, ngchmLayer, ngchmBar, or ngchmCovariate.
}
\examples{
chmColors (layer) <- c("blue", "white", "red")

}
\seealso{
\link{chmColors}
}
