% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmCovariate}
\alias{chmCovariate}
\title{Get a covariate attached to an NG-CHM dataset.}
\usage{
chmCovariate(dataset, fullname, where)
}
\arguments{
\item{dataset}{The NG-CHM dataset to get the covariate from.}

\item{fullname}{The full name of the covariate to get.
If no covariate with that name exists, return NULL.}

\item{where}{The axis or axes on which to look for the covariate  Can be "row", "column", or "both" (default).}
}
\value{
A ngchmCovariate or NULL.
}
\description{
Get a covariate attached to an NG-CHM dataset.
}
\examples{
chmCovariate (dataset, "Age")

}
\seealso{
\linkS4class{ngchmCovariate}

chmNewCovariate

chmCovariateBar
}
