% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmCovariateBar}
\alias{chmCovariateBar}
\title{Get a covariate bar attached to an NG-CHM.}
\usage{
chmCovariateBar(hm, fullname, where)
}
\arguments{
\item{hm}{The NG-CHM to get the covariate bar from.}

\item{fullname}{The full name of the covariate bar to get.
If no covariate bar with that name exists, return NULL.}

\item{where}{The axis or axes on which to look for the covariate bar  Can be "row", "column", or "both" (default).}
}
\value{
An ngchmBar or NULL.
}
\description{
Get a covariate bar attached to an NG-CHM.
}
\examples{
chmCovariateBar (hm, "Age")

}
\seealso{
\linkS4class{ngchmBar}

chmNewCovariateBar

chmCovariate
}
