\name{chmCreateServerProtocol}
\alias{chmCreateServerProtocol}
\title{Create and register an NGCHM server protocol implementation.}
\usage{
  chmCreateServerProtocol(protocolName, installMethod,
    uninstallMethod, makePrivate, makePublic)
}
\arguments{
  \item{protocolName}{The name of this protocol
  implementation.}

  \item{installMethod}{}
}
\description{
  This function creates and registers a protocol
  implementation for manipulating an NGCHM server.
}

