\name{chmDeployDir}
\alias{chmDeployDir}
\alias{chmDeployDir,ngchmServer-method}
\title{Get the deploy directory of a NGCHM server.}
\usage{
  chmDeployDir(server)
}
\arguments{
  \item{server}{The server whose deploy directory is
  required.}
}
\value{
  The path on the server to the deploy directory.
}
\description{
  Return the name of the directory on a Next Generation
  Clustered Heat Map (NGCHM) server into which a NGCHM must
  be copied for installation.
}
\seealso{
  ngchmServer-class
}

