% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmDeployServer}
\alias{chmDeployServer}
\alias{chmDeployServer,ngchmServer-method}
\title{Get the name of a NGCHM server.}
\usage{
chmDeployServer(server)

\S4method{chmDeployServer}{ngchmServer}(server)
}
\arguments{
\item{server}{The server whose name is required.}
}
\value{
The name of the server.
}
\description{
Return the name of a Next Generation Clustered Heat Map (NGCHM) server.
}
\seealso{
\linkS4class{ngchmServer}
}
