% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{chmGetDeployServerConfig}
\alias{chmGetDeployServerConfig}
\title{Get per-user configuration for a specific deploy Server.}
\usage{
chmGetDeployServerConfig(server)
}
\arguments{
\item{server}{The server for which the configuration is being set.  Must have deployServer set.}
}
\description{
Get per-user configuration for a specific deploy Server.
}

