% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmGetProperty}
\alias{chmGetProperty}
\alias{chmGetProperty,ngchmVersion2,character-method}
\alias{chmGetProperty,ngchmVersion2-method}
\title{Get a property value from an NG-CHM object}
\usage{
chmGetProperty(object, label)

\S4method{chmGetProperty}{ngchmVersion2,character}(object, label)
}
\arguments{
\item{object}{An NG-CHM object}

\item{label}{The name of the property to get}
}
\description{
Get a property value from an NG-CHM object.
}
