% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{chmGetTypeInfo}
\alias{chmGetTypeInfo}
\title{Get information about a type name.}
\usage{
chmGetTypeInfo(typename)
}
\arguments{
\item{typename}{The name of the type.}
}
\value{
A list containing basic information about the type.
}
\description{
This function gets any registered information about a type name used for determining row and column
linkouts.
Registration of a typename is (currently) not required in order to use it, so it's possible for
valid type name not to have any registered information.
}
\seealso{
chmListTypes

chmRegisterType
}

