% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmGetURL}
\alias{chmGetURL}
\alias{chmGetURL,character-method}
\alias{chmGetURL,ngchm-method}
\title{Get the URL for an installed NGCHM.}
\usage{
chmGetURL(chm, ...)

\S4method{chmGetURL}{character}(chm, server = NULL, ...)

\S4method{chmGetURL}{ngchm}(chm, server = NULL, ...)
}
\arguments{
\item{chm}{A single character string specifying the name of the NG-CHM.}

\item{...}{Ignored.}

\item{server}{The server on which to view the NGCHM}
}
\value{
A character string representing the URL of the specified NG-CHM on the specified server.
}
\description{
Return the URL for accessing the specified Next Generation Clustered Heat Map (NGCHM) on the specified server.
}
\seealso{
\linkS4class{ngchmServer}

\linkS4class{ngchm}
}
