\name{chmGetURL}
\alias{chmGetURL}
\alias{chmGetURL,ngchmServer,character-method}
\alias{chmGetURL,ngchmServer,ngchm-method}
\title{Get the URL for an installed NGCHM.}
\usage{
  chmGetURL(server, chm)
}
\arguments{
  \item{server}{The server on which the NGCHM is
  installed.}

  \item{chm}{The NGCHM for which the URL is required.}
}
\description{
  Return the URL for accessing the specified Next
  Generation Clustered Heat Map (NGCHM) on the specified
  server.
}
\seealso{
  ngchmServer-class

  ngchm-class
}

