% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\docType{methods}
\name{chmHasProperty}
\alias{chmHasProperty}
\alias{chmHasProperty,ngchmVersion2,character-method}
\alias{chmHasProperty,ngchmVersion2-method}
\title{Determine if the NG-CHM has the given property.}
\usage{
chmHasProperty(object, label)

\S4method{chmHasProperty}{ngchmVersion2,character}(object, label)
}
\arguments{
\item{object}{An NG-CHM object}
}
\description{
Determine if the NG-CHM has the given property.
}

