% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmInstall}
\alias{chmInstall}
\alias{chmInstall,ngchm-method}
\title{Add an NG-CHM to an NG-CHM collection.}
\usage{
chmInstall(chm, ...)

\S4method{chmInstall}{ngchm}(chm, path, ...)
}
\arguments{
\item{chm}{The NGCHM to install.}

\item{...}{Additional server (protocol) specific parameters.}

\item{path}{The path to the collection in which to install the NGCHM.}
}
\value{
The installed chm.
}
\description{
Add the given Next-Generation Clustered Heat Map (NG-CHM) to the specified collection (default: current collection).
}
\seealso{
\linkS4class{ngchmServer}

\linkS4class{ngchm}

\code{\link[=chmUninstall]{chmUninstall()}}

\code{\link[=chmMakePrivate]{chmMakePrivate()}}

\code{\link[=chmMakePublic]{chmMakePublic()}}
}
