% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\docType{methods}
\name{chmInstall}
\alias{chmInstall}
\alias{chmInstall,ngchm-method}
\title{Install a compiled NGCHM on a server.}
\usage{
chmInstall(chm, ...)

\S4method{chmInstall}{ngchm}(chm, server = NULL, ...)
}
\arguments{
\item{chm}{The NGCHM to install.}

\item{...}{Additional server (protocol) specific parameters.}

\item{server}{The server on which to install the NGCHM.}
}
\description{
Install the specified Next Generation Clustered Heat Map (NGCHM) on the specified server.  The installed NGCHM
is public by default.
}
\seealso{
ngchmServer-class

ngchm-class

chmUninstall

chmMakePrivate

chmMakePublic
}

