% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{chmListFunctions}
\alias{chmListFunctions}
\title{List the predefined Javascript functions available for use in NGCHM menus.}
\usage{
chmListFunctions(re = ".*")
}
\arguments{
\item{re}{Only functions with names matching re are printed (default ".*")}
}
\description{
This function lists the predefined Javascript functions available for use in NGCHM menus.
}
\seealso{
chmAddMenuItem

chmGetFunction

chmRegisterFunction

grep
}

