% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmListServers}
\alias{chmListServers}
\title{List NG-CHM Servers}
\usage{
chmListServers()
}
\value{
A character vector containing the names of all available servers.
}
\description{
This function lists all NG-CHM (Next-Generation Clustered Heat Map) servers that are currently available.
}
\examples{
servers <- chmListServers() # Get a list of all available servers.

}
