% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\docType{methods}
\name{chmLoadCHM}
\alias{chmLoadCHM}
\alias{chmLoadCHM,character,character-method}
\alias{chmLoadCHM,character,missing-method}
\alias{chmLoadCHM,ngchmServer,character-method}
\title{Load CHM from NG-CHM server}
\usage{
chmLoadCHM(serverOrURL, name)

\S4method{chmLoadCHM}{ngchmServer,character}(serverOrURL, name)

\S4method{chmLoadCHM}{character,character}(serverOrURL, name)

\S4method{chmLoadCHM}{character,missing}(serverOrURL, name)
}
\arguments{
\item{serverOrURL}{Either the server or the CHM url from which to load the CHM object.}

\item{name}{Name of the CHM to load if the first parameter is a server id}
}
\description{
Load an R CHM object from an NG-CHM server.  The CHM concerned must have been built
using this library, version 0.9.4 or later.
}

