% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmMake}
\alias{chmMake}
\alias{chmMake,ngchm-method}
\title{Compile a NGCHM.}
\usage{
chmMake(chm, ...)

\S4method{chmMake}{ngchm}(chm, ...)
}
\arguments{
\item{chm}{The NGCHM to compile.}

\item{...}{Additional chmMake options that depend on the format of the NGCHM.  For details
of the additional parameters of format x see ngchmMakeFormat.x (e.g. ngchmMakeFormat.original).}
}
\value{
The chm
}
\description{
Deprecated.  Users should no longer call this method directly.
}
\details{
Compiles the specified Next Generation Clustered Heat Map (NGCHM) in preparation for installation.
}
\seealso{
\linkS4class{ngchmServer}

\linkS4class{ngchm}

\code{\link[=chmNew]{chmNew()}}

\code{\link[=chmInstall]{chmInstall()}}

\code{\link[=ngchmMakeFormat.original]{ngchmMakeFormat.original()}}
}
