% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\docType{methods}
\name{chmMake}
\alias{chmMake}
\alias{chmMake,ngchm-method}
\title{Compile a NGCHM.}
\usage{
chmMake(chm, ...)

\S4method{chmMake}{ngchm}(chm, ...)
}
\arguments{
\item{chm}{The NGCHM to compile.}

\item{...}{Additional chmMake options that depend on the format of the NGCHM.  For details
of the additional parameters of format x see ngchmMakeFormat.x (e.g. ngchmMakeFormat.original).}
}
\value{
The chm
}
\description{
Compile the specified Next Generation Clustered Heat Map (NGCHM) in preparation for installation.
}
\seealso{
ngchmServer-class

ngchm-class

chmNew

chmInstall

ngchmMakeFormat.original
}

