\name{chmMakePublic}
\alias{chmMakePublic}
\alias{chmMakePublic,ngchmServer,character-method}
\alias{chmMakePublic,ngchmServer,ngchm-method}
\title{Make an installed NGCHM public.}
\usage{
  chmMakePublic(server, chm)
}
\arguments{
  \item{server}{The server on which the NGCHM to make
  public is installed.}

  \item{chm}{The NGCHM to make public.}
}
\description{
  Make public the specified Next Generation Clustered Heat
  Map (NGCHM) on the specified server.
}
\seealso{
  ngchmServer-class

  ngchm-class

  chmInstall

  chmUninstall

  chmMakePrivate
}

