% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmManager}
\alias{chmManager}
\title{Manage the NGCHMs on the specified server in the viewer.}
\usage{
chmManager(server = NULL, viewer = NULL)
}
\arguments{
\item{server}{The server to browse. Defaults to option "NGCHM.Server" or the first server.
The server must be managed.}

\item{viewer}{The viewer to use. Defaults to option "viewer" or browseURL.}
}
\description{
Opens the NG-CHM manager page in the viewer.
}
\seealso{
browseURL
}

