% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\docType{methods}
\name{chmName}
\alias{chmName}
\alias{chmName,ngchm-method}
\title{Get the name of a NGCHM.}
\usage{
chmName(chm)

\S4method{chmName}{ngchm}(chm)
}
\arguments{
\item{chm}{The CHM for which the name is required.}
}
\value{
A string.
}
\description{
This function returns the name of a Next Generation Clustered Heat Map (NGCHM) object.
}
\seealso{
ngchm-class
}

