% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{chmNew}
\alias{chmNew}
\title{Create a new NGCHM.}
\usage{
chmNew(name, ..., rowOrder = chmDefaultRowOrder, rowDist = "correlation",
  rowAgglom = "ward", colOrder = chmDefaultColOrder,
  colDist = "correlation", colAgglom = "ward", rowAxisType = NULL,
  colAxisType = NULL, rowCovariates = NULL, colCovariates = NULL,
  format = "original", overview = c())
}
\arguments{
\item{name}{The name under which the NGCHM will be saved to the NGCHM server.}

\item{...}{Zero or more initial objects to include in the NGCHM (see chmAdd).}

\item{rowOrder}{A vector, dendrogram, or function specifying the CHM row order.}

\item{rowDist}{Distance method to use by default RowOrder}

\item{rowAgglom}{Agglomeration method to use by default RowOrder}

\item{colOrder}{A vector, dendrogram, or function specifying the CHM column order.}

\item{colDist}{Distance method to use by default ColOrder}

\item{colAgglom}{Agglomeration method to use by default ColOrder}

\item{rowAxisType}{The type(s) of the row labels (default: None).}

\item{colAxisType}{The type(s) of the column labels (default: None).}

\item{rowCovariates}{Covariate(Bar)(s) to add to the rows (default: None).}

\item{colCovariates}{Covariate(Bar)(s) to add to the columns (default: None).}

\item{format}{The format of NGCHM to produce (default: 'original').}

\item{overview}{The format(s) of overview image(s) to create (default: None).}
}
\value{
An object of class ngchm
}
\description{
This function creates a Next Generation Clustered Heat Map (NGCHM) object in memory.
Additional parameters will be added to the new NGCHM (see chmAdd).
The bare NGCHM needs at least one data layer added to it before it can be compiled.
}
\examples{
mychm <- chmNew ("test_chm")
}
\seealso{
ngchm-class

ngchmServer-class

chmAdd

chmAddAxisType

chmAddCovariateBar

chmAddOverview

chmMake

chmInstall
}

