% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{chmNewCovariateBar}
\alias{chmNewCovariateBar}
\title{Create a new covariate Bar for a NGCHM}
\usage{
chmNewCovariateBar(covar, display = "visible", thickness = as.integer(10),
  merge = NULL)
}
\arguments{
\item{covar}{The covariate to be displayed in the bar.}

\item{display}{Whether the covariate bar will be "hidden" or "visible" (default).}

\item{thickness}{The thickness of the covariate bar in pixels. (Default 10).}

\item{merge}{Algorithm for merging covariates when necessary ("average", "peakColor", "specialColor", or "mostCommon").}
}
\value{
An object of class ngchmBar
}
\description{
This function creates a new covariate bar suitable for adding to a Next Generation Clustered Heat Map.
}
\seealso{
ngchmBar-class

chmNewColorMap

chmAddCovariateBar
}

