% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{chmNewServer}
\alias{chmNewServer}
\title{Create a new object representing a NGCHM server.}
\usage{
chmNewServer(serverName, username = NULL, keypath = NULL,
  serverPort = 8080, deployServer = NULL, protoOpts = NULL,
  jarFile = NULL, urlBase = NULL)
}
\arguments{
\item{serverName}{The DNS name of the NGCHM server.}

\item{username}{The login name to use when deploying to the server (defaults to the username ssh uses).}

\item{keypath}{The path to the key to use when deploying to the server (defaults to your normal ssh credentials).}

\item{serverPort}{The port on which the server is listening.}

\item{deployServer}{The DNS name to use when deploying a NGCHM (defaults to serverName).}

\item{protoOpts}{A list of protocol-specific parameters}

\item{jarFile}{The location of the heatmap build jar file to use when making a NGCHM (defaults to jar file on urlBase WS).}

\item{urlBase}{The base URL used to access a deployed NGCHM (defaults to serverName:serverPort/chm/chm.html).}
}
\value{
An object of class ngchmServer
}
\description{
This function creates a new object that represents a NGCHM server.
}
\examples{
cloudServ <- chmNewServer ("dnsname.domain")
}
\seealso{
ngchmServer-class

chmMake

chmInstall

chmUninstall
}

