% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmProperties}
\alias{chmProperties}
\title{Create a list of NGCHM properties.}
\usage{
chmProperties(...)
}
\arguments{
\item{...}{A named list of property values.}
}
\description{
Create a list of NGCHM properties.
}
\seealso{
\code{\link[=chmAdd]{chmAdd()}}
}
