% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmRegisterFunction}
\alias{chmRegisterFunction}
\title{Register a predefined Javascript function for use in NGCHM menus.}
\usage{
chmRegisterFunction(fn)
}
\arguments{
\item{fn}{The function to be registered. This should be an object of class 'ngchmJS'.}
}
\value{
The registered function.
}
\description{
This function registers a Javascript function that can be used when
building a Next Generation Clustered Heat Map.  This function is
intended for use by NGCHM system developers.
}
\examples{
# Assume 'myFunction' is an object of class 'ngchmJS'
chmRegisterFunction(myFunction) # Register 'myFunction'.

}
\seealso{
\code{\link[=chmAddMenuItem]{chmAddMenuItem()}}

\code{\link[=chmNewFunction]{chmNewFunction()}}

\linkS4class{ngchmAxisFunction}

\linkS4class{ngchmMatrixFunction}
}
