% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{chmRegisterFunction}
\alias{chmRegisterFunction}
\title{Register a predefined Javascript function for use in NGCHM menus.}
\usage{
chmRegisterFunction(fn)
}
\arguments{
\item{fn}{The Javascript function to register.}
}
\description{
This function registers a Javascript function that can be used when
building a Next Generation Clustered Heat Map.  This function is
intended for use by NGCHM system developers.
}
\seealso{
chmAddMenuItem

chmNewFunction

ngchmFunction-class
}

