% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{chmRegisterGetMetadataFunction}
\alias{chmRegisterGetMetadataFunction}
\title{Define and register a Javascript function for obtaining a specific metadata value.}
\usage{
chmRegisterGetMetadataFunction(functionName, metadataColumnName)
}
\arguments{
\item{functionName}{The name given to the Javascript function.}

\item{metadataColumnName}{The name of the metadata column to access.}
}
\description{
This function defines and registers a Javascript function for obtaining a specific
metadata value and returning it as a javascript list.  The function is suitable for
use as an axis type accessor function.
}
\seealso{
chmAddAxisType

chmGetFunction

chmListFunctions
}

