% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmRegisterType}
\alias{chmRegisterType}
\title{Register a type name.}
\usage{
chmRegisterType(typename, description)
}
\arguments{
\item{typename}{A character vector specifying the name(s) of the type(s) to be
registered.}

\item{description}{A single character string specifying the description of the
type(s).}
}
\value{
None. This function is used for its side effects of registering a new type.
}
\description{
This function registers a type name used for determining row and column
linkouts.
This function is intended to be used by NGCHM system developers to record
basic information about the semantic interpretation of a type name.  Registration
of a typename is (currently) not required in order to use it.
}
\seealso{
\code{\link[=chmListTypes]{chmListTypes()}}

\code{\link[=chmGetTypeInfo]{chmGetTypeInfo()}}

\code{\link[=chmRegisterTypeMapper]{chmRegisterTypeMapper()}}
}
