% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\docType{methods}
\name{chmRowOrder<-}
\alias{chmRowOrder<-}
\alias{chmRowOrder<-,ngchm,optDendrogram-method}
\title{Set the row order of data shown in a NGCHM.}
\usage{
chmRowOrder(chm) <- value

\S4method{chmRowOrder}{ngchm,optDendrogram}(chm) <- value
}
\arguments{
\item{chm}{The chm for which to set the row order.}

\item{value}{An ordering of the row labels.
If value is NULL, the labels will be displayed in the same order they are found in the first data layer.
If value is a character vector, the labels will be displayed in that order.
If value is a dendrogram, the labels displayed in the order they occur in a depth first traversal of the tree.}
}
\description{
Set the order in which rows in the data will be shown in a Next Generation Clustered Heat Map.
}
\seealso{
"chmColOrder<-"
}

