% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmSetCredentials}
\alias{chmSetCredentials}
\alias{chmSetCredentials,ngchmServer,character-method}
\alias{chmSetCredentials,character,character-method}
\title{Set access credentials.}
\usage{
chmSetCredentials(resource, credentials)

\S4method{chmSetCredentials}{ngchmServer,character}(resource, credentials)

\S4method{chmSetCredentials}{character,character}(resource, credentials)
}
\arguments{
\item{resource}{The resource (e.g. ngchmServer) to which the credentials apply.}

\item{credentials}{The credentials.}
}
\description{
Set access credentials for the specific resource.
}
