% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{chmSetDeployServerConfig}
\alias{chmSetDeployServerConfig}
\title{Specify per-user configuration for a specific deploy Server.}
\usage{
chmSetDeployServerConfig(server, config)
}
\arguments{
\item{server}{The server for which the configuration is being set.  Must have deployServer set.}

\item{config}{The configuration to set}
}
\description{
Specify per-user configuration for a specific deploy Server.
}
