% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{chmSetDisplayLength}
\alias{chmSetDisplayLength}
\alias{chmSetDisplayLength,ngchmVersion2-method}
\title{Set number of characters to display for row or column labels}
\usage{
chmSetDisplayLength(object, value, rowOrCol)

\S4method{chmSetDisplayLength}{ngchmVersion2}(object, value, rowOrCol)
}
\arguments{
\item{object}{An ngchmVersion2 object to modify}

\item{value}{Numeric value to set as display length}

\item{rowOrCol}{Character string indicating which labels to modify: "row" or "col" ("column" also accepted)}
}
\value{
Modified ngchmVersion2 object with updated display length
}
\description{
Sets the maximum number of characters to display for row or column labels in the NG-CHM viewer.
Allowed values are: 10, 15, 20, 25, 30, 35, 40.
}
\examples{
# Set row labels to display up to 20 characters
chm <- chmNew("New Heat Map")
chm <- chmSetDisplayLength(chm, 20, "row")

# Set column labels to display up to 15 characters
chm <- chmSetDisplayLength(chm, 15, "col")

}
