% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmTreeGaps}
\alias{chmTreeGaps}
\title{Creates new treeCuts object}
\usage{
chmTreeGaps(numberOfCuts)
}
\arguments{
\item{numberOfCuts}{Number of tree cuts}
}
\value{
\linkS4class{treeCuts} object with specified number of tree cuts
}
\description{
This function was designed to facilitate setting rowGapLocations and colGapLocations in the
\code{\link[=chmNew]{chmNew()}} function. See examples section.
}
\examples{
mychm <- chmNew("test_chm", rowGapLocations = chmTreeGaps(5))
}
