% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\docType{methods}
\name{chmUninstall}
\alias{chmUninstall}
\alias{chmUninstall,character-method}
\alias{chmUninstall,ngchm-method}
\title{Remove an installed NGCHM.}
\usage{
chmUninstall(chm, ...)

\S4method{chmUninstall}{character}(chm, server = NULL, ...)

\S4method{chmUninstall}{ngchm}(chm, ...)
}
\arguments{
\item{chm}{The NGCHM to remove.}

\item{...}{Additional server (protocol) specific parameters.}

\item{server}{The server on which the NGCHM to remove is installed.}
}
\description{
Remove the specified Next Generation Clustered Heat Map (NGCHM) from the specified server.
}
\seealso{
ngchmServer-class

ngchm-class

chmInstall
}

