% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\docType{methods}
\name{chmUninstall}
\alias{chmUninstall}
\alias{chmUninstall,character,character-method}
\alias{chmUninstall,character,ngchm-method}
\alias{chmUninstall,ngchmServer,character-method}
\alias{chmUninstall,ngchmServer,ngchm-method}
\title{Remove an installed NGCHM.}
\usage{
chmUninstall(server, chm, ...)

\S4method{chmUninstall}{ngchmServer,character}(server, chm, ...)

\S4method{chmUninstall}{ngchmServer,ngchm}(server, chm, ...)

\S4method{chmUninstall}{character,ngchm}(server, chm, ...)

\S4method{chmUninstall}{character,character}(server, chm, ...)
}
\arguments{
\item{server}{The server on which the NGCHM to remove is installed.}

\item{chm}{The NGCHM to remove.}

\item{...}{Additional server (protocol) specific parameters.}
}
\description{
Remove the specified Next Generation Clustered Heat Map (NGCHM) from the specified server.
}
\seealso{
ngchmServer-class

ngchm-class

chmInstall
}

