% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\docType{methods}
\name{chmUrlBase}
\alias{chmUrlBase}
\alias{chmUrlBase,ngchmServer-method}
\title{Get the base URL for a NGCHM installed on a NGCHM server.}
\usage{
chmUrlBase(server)

\S4method{chmUrlBase}{ngchmServer}(server)
}
\arguments{
\item{server}{The server whose base URL is required.}
}
\value{
The base URL for accessing NGCHMs installed on the server.
}
\description{
Return the base URL of a Next Generation Clustered Heat Map (NGCHM) that
has been installed on a NGCHM server.
}
\seealso{
ngchmServer-class
}

