% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmUserDendrogramToShaid}
\alias{chmUserDendrogramToShaid}
\title{Convert a user specified dendrogram to a shaid}
\usage{
chmUserDendrogramToShaid(ddg)
}
\arguments{
\item{ddg}{The dendrogram to convert.}
}
\description{
Convert a user specified dendrogram to a shaid
}

