% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmUserLabelsToShaid}
\alias{chmUserLabelsToShaid}
\title{Convert user specified labels to a shaid}
\usage{
chmUserLabelsToShaid(labels)
}
\arguments{
\item{labels}{A string vector containing axis labels.}
}
\description{
Convert user specified labels to a shaid
}
