% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chmWriteCustomJS}
\alias{chmWriteCustomJS}
\title{Output Javascript code required to customize an NGCHM.}
\usage{
chmWriteCustomJS(chm, filename)
}
\arguments{
\item{chm}{An NGCHM}

\item{filename}{The file to which the customization code will be written.}
}
\description{
This function outputs the Javascript required to customize an NGCHM.
}

