% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{initLogging}
\alias{initLogging}
\title{Function to initialize logging}
\usage{
initLogging(log_level, log_file = NULL)
}
\arguments{
\item{log_level}{One of 'TRACE', 'DEBUG', 'INFO', 'SUCCESS'}

\item{log_file}{Desired path/name of log file}
}
\description{
Function to initialize logging
}
