\name{ngchm-class}
\alias{ngchm-class}
\title{Class representing a Next Generation Clustered Heat Map (NGCHM) under construction.}
\description{
  Class representing a Next Generation Clustered Heat Map
  (NGCHM) under construction.
}
\seealso{
  chmNew

  chmRowOrder

  chmColOrder

  chmAddLayer

  chmAddClassBar

  chmAddDataset

  chmAddMenuItem

  chmAddCSS

  chmMake

  chmInstall
}
\keyword{classes}

