% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allClasses.R
\name{ngchm-class}
\alias{ngchm-class}
\title{Class representing a Next Generation Clustered Heat Map (NGCHM) under construction.}
\description{
Class representing a Next Generation Clustered Heat Map (NGCHM) under construction.
}
\seealso{
chmNew

chmRowOrder

chmColOrder

chmAddLayer

chmAddCovariateBar

chmAddDataset

chmAddMenuItem

chmAddCSS

chmMake

chmInstall
}
\keyword{classes}

