\name{NGCHM-functions}
\alias{NGCHM-functions}
\title{Javascript extensions for the Next Generation Clustered Heat Map (NGCHM) Construction Library}
\description{
  Currently: \itemize{ \item Axis function View Ideogram is
  added for the appropriate axis types. }
}
\seealso{
  chmGetFunction

  chmListFunctions
}

