% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{NGCHM}
\alias{NGCHM}
\alias{NGCHM-package}
\title{Next Generation Clustered Heat Map (NGCHM) Construction Library}
\description{
NGCHM provides tools for defining the contents of a new NGCHM,
and for compiling and installing it on a NGCHM server.
}
\details{
Typical usage (see example) is to create a base NGCHM using chmNew; extend it with at
least one ngchmLayer; typically extend it further with an additional
ngchmLayer, row and column dendrograms, classification bars, and popup menu entries;
compile and install it on an available ngchmServer.
}
\section{Initialization}{

When first loaded the NGCHM library reads configuration files in
the directories specified by the NGCHMCONFIGPATH environment variable.  This is
a colon (:) separated list of directory names.  If not set it defaults to
/etc/ngchm:/usr/local/ngchm:/opt/ngchm:$HOME/.ngchm.
See NGCHM-initialization for details.
}
\examples{
chm <- chmNew ('my-chm', mat, rowAxisType='bio.gene.hugo')
chmMake ('my-server', chm)
chmInstall ('my-server', chm)
}
\seealso{
chmNew

chmAdd

chmMake

chmInstall

chmListServers

ngchm-class
}

