% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmAddDatasetBlob}
\alias{ngchmAddDatasetBlob}
\title{Add a data file to a local shaidy repository}
\usage{
ngchmAddDatasetBlob(shaidyRepo, format, filename, properties = NULL)
}
\arguments{
\item{shaidyRepo}{The shaidy repository}

\item{format}{The format of the data file}

\item{filename}{The filesystem path to the data file}

\item{properties}{A list of additional properties to save with file}
}
\value{
The file's shaid
}
\description{
Add a data file to a local shaidy repository
}
