% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmAddObjectToCollection}
\alias{ngchmAddObjectToCollection}
\title{Add an object reference to a collection}
\usage{
ngchmAddObjectToCollection(repo, uuid, shaid)
}
\arguments{
\item{repo}{The repository containing the collection}

\item{uuid}{A collection uuid}

\item{shaid}{The shaid of the object to add to the collection}
}
\value{
An updated list containing details of the collection
}
\description{
Add an object reference to a collection
}

