% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ngchmAxis-class}
\alias{ngchmAxis-class}
\title{Class representing an axis of a Next Generation Clustered Heat Map (NG-CHM).}
\description{
Class representing an axis of a Next Generation Clustered Heat Map (NG-CHM).
}
\seealso{
\code{\link[=chmAxis]{chmAxis()}}
}
\keyword{classes}
