% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ngchmAxisFunction-class}
\alias{ngchmAxisFunction-class}
\title{Class representing an axis function for Next Generation Clustered Heat Map (NGCHM).}
\description{
Class representing an axis function for Next Generation Clustered Heat Map (NGCHM).
}
\keyword{classes}
