% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ngchmCreateServerProtocol}
\alias{ngchmCreateServerProtocol}
\title{Create and register an NGCHM server protocol implementation.}
\usage{
ngchmCreateServerProtocol(
  protocolName,
  chmFormat,
  requiredParams,
  optionalParams,
  paramValidator,
  findCollection,
  createCollection,
  installMethod,
  uninstallMethod,
  makePrivate,
  makePublic,
  setCredentials
)
}
\arguments{
\item{protocolName}{The name of this protocol implementation.}

\item{chmFormat}{The chm format required by this protocol. Defaults to original format.}

\item{requiredParams}{The protocol's required parameters, if any.}

\item{optionalParams}{The protocol's optional parameters, if any.}

\item{paramValidator}{A function(list) for validating the parameters specified for a new server.}

\item{findCollection}{A function(server,collection,path) for finding a collection.}

\item{createCollection}{A function(server,collection,name) for creating a collection.}

\item{installMethod}{A function(server,chm) for installing an NG-CHM.}

\item{uninstallMethod}{A function(server,chmname) for uninstalling an NG-CHM.}

\item{makePrivate}{A function(server,chmname) for hiding an NG-CHM.}

\item{makePublic}{A function(server,chmname) for showing an NG-CHM.}

\item{setCredentials}{A function(server,credentialstring) for setting credentials.}
}
\description{
This function creates and registers a protocol implementation for manipulating
an NGCHM server.
}
