% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{ngchmCreateServerProtocol}
\alias{ngchmCreateServerProtocol}
\title{Create and register an NGCHM server protocol implementation.}
\usage{
ngchmCreateServerProtocol(protocolName, requiredParams, optionalParams,
  paramValidator, findCollection, createCollection, installMethod,
  uninstallMethod, makePrivate, makePublic)
}
\arguments{
\item{protocolName}{The name of this protocol implementation.}

\item{requiredParams}{The protocol's required parameters, if any.}

\item{optionalParams}{The protocol's optional parameters, if any.}

\item{paramValidator}{A function(list) for validating the parameters specified for a new server.}

\item{findCollection}{A function(server,collection,path) for finding a collection.}

\item{createCollection}{A function(server,collection,name) for creating a collection.}

\item{installMethod}{A function(server,chm) for installing an NG-CHM.}

\item{uninstallMethod}{A function(server,chmname) for uninstalling an NG-CHM.}

\item{makePrivate}{A function(server,chmname) for hiding an NG-CHM.}

\item{makePublic}{A function(server,chmname) for showing an NG-CHM.}
}
\description{
This function creates and registers a protocol implementation for manipulating
an NGCHM server.
}

