% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ngchmCSS-class}
\alias{ngchmCSS-class}
\title{Class representing custom CSS for a Next Generation Clustered Heat Map (NGCHM).}
\description{
Class representing custom CSS for a Next Generation Clustered Heat Map (NGCHM).
}
\keyword{classes}
