% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmGetLabels}
\alias{ngchmGetLabels}
\title{Get the axis labels of a shaidy dataset or dendrogram}
\usage{
ngchmGetLabels(shaid, axis = NULL)
}
\arguments{
\item{shaid}{The shaid of the dataset or dendrogram to get the labels of}

\item{axis}{For datasets, the axis of the labels to get}
}
\value{
a list of shaids containing the labels
}
\description{
Get the axis labels of a shaidy dataset or dendrogram
}

