% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{ngchmGetProtoParam}
\alias{ngchmGetProtoParam}
\title{Get a server protocol parameter}
\usage{
ngchmGetProtoParam(server, option, default = NULL)
}
\arguments{
\item{server}{The NGCHM server}

\item{option}{The name of the protocol parameter to get}

\item{default}{The default value of the option (default NULL)}
}
\description{
Get a server protocol parameter
}

