% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ngchmGetServerProtocol}
\alias{ngchmGetServerProtocol}
\title{Get Server Protocol for NG-CHM}
\usage{
ngchmGetServerProtocol(protocolName)
}
\arguments{
\item{protocolName}{A single character string specifying the name of the protocol.}
}
\value{
An object of class 'ngchmServerProtocol' representing the server protocol.
}
\description{
This function gets a server protocol for NG-CHM (Next-Generation Clustered Heat Map)
by its name.
}
