% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmInitShaidyRepository}
\alias{ngchmInitShaidyRepository}
\title{Initialize Shaidy Repository for NG-CHM}
\usage{
ngchmInitShaidyRepository(shaidyDir)
}
\arguments{
\item{shaidyDir}{A single character string specifying the directory where the
Shaidy repository will be initialized.}
}
\value{
None. This function is used for its side effects of initializing the
Shaidy repository.
}
\description{
This function initializes a Shaidy repository for NG-CHM (Next-Generation Clustered Heat Map).
}
