% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ngchmListServerProtocols}
\alias{ngchmListServerProtocols}
\title{List defined server protocols}
\usage{
ngchmListServerProtocols()
}
\value{
A character vector
}
\description{
List defined server protocols
}

