% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmLoadDatasetBlob}
\alias{ngchmLoadDatasetBlob}
\title{Load a data matrix from a local shaidy repository}
\usage{
ngchmLoadDatasetBlob(shaidyRepo, shaid, datatype)
}
\arguments{
\item{shaidyRepo}{The shaidy repository}

\item{shaid}{The shaid of the dataset blob to load}

\item{datatype}{Prototype of matrix data elements (defaults to 0.0)}
}
\value{
a list containing details of the loaded dataset
}
\description{
Load a data matrix from a local shaidy repository
}
