% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ngchmProtoParamCheck}
\alias{ngchmProtoParamCheck}
\title{Check that all required parameters are specified, and all specified parameters are either required or optional.}
\usage{
ngchmProtoParamCheck(params, required, optional)
}
\arguments{
\item{params}{A list of named parameters.}

\item{required}{A character vector of required parameter names}

\item{optional}{A character vector of optional parameter names}
}
\description{
Check that all required parameters are specified, and all specified parameters are either required or optional.
}
