% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmPushSourceRepository}
\alias{ngchmPushSourceRepository}
\title{Push a local shaidy repository onto the stack of source repositories}
\usage{
ngchmPushSourceRepository(shaidyDir, accessMethod = "file")
}
\arguments{
\item{shaidyDir}{A single character string specifying the directory of the
source repository.}

\item{accessMethod}{A single character string specifying the access method
for the source repository. Defaults to "file".}
}
\value{
None. This function is used for its side effects of pushing the source
repository onto the Shaidy stack.
}
\description{
This function pushes a source repository for NG-CHM (Next-Generation
Clustered Heat Map) onto the Shaidy stack.
}
