% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmPushSourceServer}
\alias{ngchmPushSourceServer}
\title{Push a shaidy server onto the stack of source repositories}
\usage{
ngchmPushSourceServer(server)
}
\arguments{
\item{server}{An ngchmServer or the name of one.}
}
\description{
Push a shaidy server onto the stack of source repositories
}
\seealso{
\code{\link[=chmLoadShaidyCHM]{chmLoadShaidyCHM()}}

\code{\link[=chmCreateServer]{chmCreateServer()}}
}
