% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ngchmshaidy.R
\name{ngchmPushTempRepository}
\alias{ngchmPushTempRepository}
\title{Push a shaidy repository onto the stack of temporary repositories}
\usage{
ngchmPushTempRepository(shaidyDir)
}
\arguments{
\item{shaidyDir}{Basepath of local shaidy repository to use as a temporary repository}
}
\description{
Push a shaidy repository onto the stack of temporary repositories
}

