\name{ngchmServerProtocol-class}
\alias{ngchmServerProtocol-class}
\title{Class representing a deployment method for a Next Generation Clustered Heat Map (NGCHM) server.}
\description{
  Class representing a deployment method for a Next
  Generation Clustered Heat Map (NGCHM) server.
}
\keyword{classes}

