% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.ngchm.type.info}
\alias{print.ngchm.type.info}
\title{Pretty Print NGCHM Type Information}
\usage{
\method{print}{ngchm.type.info}(x, ...)
}
\arguments{
\item{x}{An object of class 'ngchm.type.info' as returned by chmGetTypeInfo.}

\item{...}{Additional arguments (not used).}
}
\value{
A string that provides a detailed description of the NGCHM type.
}
\description{
This function takes an object of class 'ngchm.type.info' and returns a formatted string
that provides a detailed description of the NGCHM type.
}
\examples{
# Assume 'typeInfo' is an object of class 'ngchm.type.info'
print.ngchm.type.info(typeInfo)

}
\seealso{
\code{\link[=chmGetTypeInfo]{chmGetTypeInfo()}}
}
