% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.ngchm.type.info}
\alias{print.ngchm.type.info}
\title{Pretty print the result returned by chmGetTypeInfo.}
\usage{
\method{print}{ngchm.type.info}(x, ...)
}
\arguments{
\item{x}{Type information about an NGCHM type name as return by chmGetTypeInfo.}

\item{...}{Generic function accepts additional parameters}
}
\description{
Pretty print the result returned by chmGetTypeInfo.
}
\seealso{
chmGetTypeInfo
}

