% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaidy.R
\name{print.shaidyRepo}
\alias{print.shaidyRepo}
\title{Print a shaidy repository}
\usage{
\method{print}{shaidyRepo}(x, ...)
}
\arguments{
\item{x}{The shaidy repository to print}

\item{...}{Unused extra parameters}
}
\value{
The shaidy repository
}
\description{
Print a shaidy repository
}

