% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaidy.R
\name{shaidyAddFileBlob}
\alias{shaidyAddFileBlob}
\title{Add data file(s) and properties to a local shaidy repository}
\usage{
shaidyAddFileBlob(
  shaidyRepo,
  blob.type,
  blob.file,
  filename,
  properties = NULL,
  shaid = NULL
)
}
\arguments{
\item{shaidyRepo}{A shaidy repository object}

\item{blob.type}{The type of the data file(s) (e.g. 'dataset', 'label', 'chm')}

\item{blob.file}{Name(s) of the file(s) within the blob}

\item{filename}{The filesystem path(s) to the file(s) to insert}

\item{properties}{A list of optional additional properties to save with the file(s)}

\item{shaid}{Shaid to store the blob as.}
}
\value{
The file's shaid
}
\description{
This function adds a blob (file(s)) of a specified type to a Shaidy repository.
The file blob is created in a temporary directory in the blob path of the repository.
If a blob with the same shaid already exists in the repository, the file blob is not added.
}
