% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shaidy.R
\name{shaidyAddFileBlob}
\alias{shaidyAddFileBlob}
\title{Add a data file to a local shaidy repository}
\usage{
shaidyAddFileBlob(shaidyRepo, blob.type, blob.file, filename,
  properties = NULL, shaid = NULL)
}
\arguments{
\item{shaidyRepo}{The shaidy repository}

\item{blob.type}{The blob.type of the data file}

\item{blob.file}{Name of file within the blob}

\item{filename}{The filesystem path to the file to insert}

\item{properties}{A list of additional properties to save with file}

\item{shaid}{Shaid to store the blob as.}
}
\value{
The file's shaid
}
\description{
Add a data file to a local shaidy repository
}

