% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shaidy.R
\name{shaidyBlobPath}
\alias{shaidyBlobPath}
\title{Return blob.path function for a local shaidy repository.}
\usage{
shaidyBlobPath(shaidyDir)
}
\arguments{
\item{shaidyDir}{Basepath to a local shaidy repository.}
}
\value{
a function (first, ...) that accepts either a shaid or a blob type and optionally
        additional file path components and returns a filepath
}
\description{
Return blob.path function for a local shaidy repository.
}

