% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaidy.R
\name{shaidyCreateProtoBlob}
\alias{shaidyCreateProtoBlob}
\title{Create a prototype blob in a shaidy repository}
\usage{
shaidyCreateProtoBlob(shaidyRepo, blob.type)
}
\arguments{
\item{shaidyRepo}{The shaidy repository}

\item{blob.type}{The blob.type of the prototype blob}
}
\value{
The file path of the prototype blob
}
\description{
Create a prototype blob in a shaidy repository
}
