% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaidy.R
\name{shaidyFinalizeProtoBlob}
\alias{shaidyFinalizeProtoBlob}
\title{Finalize a prototype blob}
\usage{
shaidyFinalizeProtoBlob(shaidyRepo, shaid, protoblob)
}
\arguments{
\item{shaidyRepo}{The shaidy repository}

\item{shaid}{The shaid to assign the protoblob}

\item{protoblob}{The prototype blob to finalize}
}
\value{
The shaid (invisibly)

The protoblob must have been created in the specified shaidy repository
and with the same blob type as the shaid.  When this function returns the
protoblob will no longer be accessible .  If a blob with the same shaid already
exists in this repository, the protoblob is quitely removed without affecting
the existing blob.
}
\description{
Finalize a prototype blob
}
