% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allMethods.R
\name{shaidyGetComponents}
\alias{shaidyGetComponents}
\alias{shaidyGetComponents,ngchm-method}
\alias{shaidyGetComponents,ngchmDataset-method}
\alias{shaidyGetComponents,ngchmCovariate-method}
\title{Get an object's component shaids}
\usage{
shaidyGetComponents(object)

\S4method{shaidyGetComponents}{ngchm}(object)

\S4method{shaidyGetComponents}{ngchmDataset}(object)

\S4method{shaidyGetComponents}{ngchmCovariate}(object)
}
\arguments{
\item{object}{The object, such as a chm, dataset, etc., for which to get the component shaids}
}
\value{
A list of shaids.
}
\description{
Get an object's component shaids
}
