% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaidy.R
\name{shaidyInitRepository}
\alias{shaidyInitRepository}
\title{Create and initialize a local shaidy repository.}
\usage{
shaidyInitRepository(shaidyDir, blob.types)
}
\arguments{
\item{shaidyDir}{Basepath of local shaidy repository to create}

\item{blob.types}{A string vector naming the blob types to include in the repository}
}
\description{
Create and initialize a local shaidy repository.
}
