% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shaidy.R
\name{shaidyLoadRepository}
\alias{shaidyLoadRepository}
\title{Load a local shaidy repository}
\usage{
shaidyLoadRepository(shaidyDir)
}
\arguments{
\item{shaidyDir}{Basepath to a local shaidy repository.}
}
\value{
A shaidyRepo
}
\description{
Load a local shaidy repository
}

