% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{treeCuts-class}
\alias{treeCuts-class}
\title{Helper class for setting row/col gap locations as tree cuts}
\description{
This class is to facilitate specification of row/col gaps in \code{\link[=chmNew]{chmNew()}}.
Note: user-facing function use the term 'gap', while internal functions that
interact with java programs in the NGCHM viewer project use the term 'cut'.
}
\section{Slots}{

\describe{
\item{\code{numberOfCuts}}{Integer number of cuts}
}}

\seealso{
\code{\link[=chmNew]{chmNew()}}

\code{\link[=chmTreeGaps]{chmTreeGaps()}}
}
