% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{verifyNumeric}
\alias{verifyNumeric}
\title{Helper function to verify if variable is numeric.}
\usage{
verifyNumeric(variableToCheck)
}
\arguments{
\item{variableToCheck}{The variable to check for being numeric.}
}
\value{
TRUE
}
\description{
If not numeric, print error message and stop.
}
